% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{add_diet_data}
\alias{add_diet_data}
\title{Agregar datos de dieta a objeto Bioenergetic}
\usage{
add_diet_data(bio_obj, diet_data)
}
\arguments{
\item{bio_obj}{Objeto Bioenergetic}

\item{diet_data}{Data frame con proporciones de dieta}
}
\value{
Objeto Bioenergetic modificado
}
\description{
Agregar datos de dieta a objeto Bioenergetic
}
