% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{add_temperature_data}
\alias{add_temperature_data}
\title{Agregar datos de temperatura a objeto Bioenergetic}
\usage{
add_temperature_data(bio_obj, temperature_data)
}
\arguments{
\item{bio_obj}{Objeto Bioenergetic}

\item{temperature_data}{Data frame con columnas Day y Temperature}
}
\value{
Objeto Bioenergetic modificado
}
\description{
Agregar datos de temperatura a objeto Bioenergetic
}
