% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{analyze_fb4_statistics}
\alias{analyze_fb4_statistics}
\title{Análisis estadístico de resultados FB4}
\usage{
analyze_fb4_statistics(fb4_result)
}
\arguments{
\item{fb4_result}{Objeto fb4_result}
}
\value{
Lista con estadísticas descriptivas
}
\description{
Análisis estadístico de resultados FB4
}
