% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{Ash}
\alias{Ash}
\title{Estimar contenido de ceniza a partir del agua}
\usage{
Ash(H2O)
}
\arguments{
\item{H2O}{Contenido de agua (g)}
}
\value{
Contenido de ceniza (g)
}
\description{
Estima gramos de ceniza basado en el contenido de agua usando la
regresión de Breck (2014)
}
\examples{
# Para un pez de 100g con 72\% de agua
water_content <- 100 * 0.72
ash_content <- Ash(water_content)
print(ash_content)
}
\references{
Breck, J.E. 2014. Body composition in fishes: body size matters.
Aquaculture 433:40-49.
}
