% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{binary_search_p_value_robust}
\alias{binary_search_p_value_robust}
\title{Robust Binary Search for P-value (Improved Version)}
\usage{
binary_search_p_value_robust(
  target_value,
  fit_type,
  species_params,
  initial_weight,
  processed_data,
  model_options,
  oxycal = 13560,
  tolerance = 0.001,
  max_iterations = 25
)
}
\arguments{
\item{target_value}{Target value (final weight or total consumption)}

\item{fit_type}{Type of fitting ("weight" or "consumption")}

\item{species_params}{Species parameters}

\item{initial_weight}{Initial weight}

\item{processed_data}{Processed data}

\item{model_options}{Model options}

\item{oxycal}{Oxycalorific coefficient}

\item{tolerance}{Tolerance for convergence}

\item{max_iterations}{Maximum number of iterations}
}
\value{
List with fitted p-value and fitting information
}
\description{
Improved version based on critical function from first script
}
