% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_activity_factor}
\alias{calculate_activity_factor}
\title{Función de actividad para respiración - Ecuación 1}
\usage{
calculate_activity_factor(
  weight,
  temperature,
  RTL,
  ACT,
  RK4,
  BACT,
  RK1,
  RK5,
  RTO
)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{temperature}{Temperatura del agua (°C)}

\item{RTL}{Temperatura de corte para cambio en relación de actividad}

\item{ACT}{Multiplicador de actividad}

\item{RK4}{Coeficiente de dependencia de masa para velocidad de nado}

\item{BACT}{Coeficiente de dependencia de temperatura para velocidad de nado}

\item{RK1}{Intercepto para velocidad de nado sobre temperatura de corte}

\item{RK5}{Coeficiente de temperatura para velocidad de nado}

\item{RTO}{Temperatura óptima para respiración}
}
\value{
Factor de actividad
}
\description{
Implementa la función de actividad con componente de temperatura
}
\keyword{internal}
