% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_activity_factor_respiration}
\alias{calculate_activity_factor_respiration}
\title{Calculate activity factor for respiration (Mid-level)}
\usage{
calculate_activity_factor_respiration(
  weight,
  temperature,
  processed_respiration_params
)
}
\arguments{
\item{weight}{Fish weight (g)}

\item{temperature}{Water temperature (°C)}

\item{processed_respiration_params}{List with processed respiration parameters (includes activity data)}
}
\value{
Activity factor
}
\description{
Calculates activity factor based on respiration equation with support for
both simple and complex activity calculations
}
\keyword{internal}
