% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_assimilation_efficiency}
\alias{calculate_assimilation_efficiency}
\title{Calcular eficiencia de asimilación}
\usage{
calculate_assimilation_efficiency(C, Eg)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Eg}{Egestion (J/g)}
}
\value{
Eficiencia de asimilación (fracción)
}
\description{
Calcula la eficiencia con la que el alimento es asimilado
}
