% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{calculate_component_from_water}
\alias{calculate_component_from_water}
\title{Estimate composition from water using Breck (2014) regression (Low-level)}
\usage{
calculate_component_from_water(water_content, component_type)
}
\arguments{
\item{water_content}{Water content (g)}

\item{component_type}{Type of component ("protein" or "ash")}
}
\value{
Component content (g)
}
\description{
Estimates grams of component based on water content using Breck regression
}
\references{
Breck, J.E. 2014. Body composition in fishes: body size matters. Aquaculture 433:40-49.
}
\keyword{internal}
