% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{calculate_consumption_params_eq2}
\alias{calculate_consumption_params_eq2}
\title{Calcular parámetros adicionales para ecuación de consumo 2}
\usage{
calculate_consumption_params_eq2(CQ, CTM, CTO)
}
\arguments{
\item{CQ}{Coeficiente de temperatura}

\item{CTM}{Temperatura máxima}

\item{CTO}{Temperatura óptima}
}
\value{
Lista con CY, CZ, CX
}
\description{
Calcular parámetros adicionales para ecuación de consumo 2
}
