% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{calculate_consumption_params_eq3}
\alias{calculate_consumption_params_eq3}
\title{Calcular parámetros adicionales para ecuación de consumo 3}
\usage{
calculate_consumption_params_eq3(CTO, CQ, CTL, CTM, CK1, CK4)
}
\arguments{
\item{CTO}{Temperatura óptima}

\item{CQ}{Coeficiente de temperatura}

\item{CTL}{Temperatura con tasa reducida}

\item{CTM}{Temperatura máxima}

\item{CK1}{Fracción pequeña de tasa máxima}

\item{CK4}{Fracción reducida de tasa máxima}
}
\value{
Lista con CG1, CG2
}
\description{
Calcular parámetros adicionales para ecuación de consumo 3
}
