% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{calculate_cumulative_survival}
\alias{calculate_cumulative_survival}
\title{Calcular supervivencia acumulativa}
\usage{
calculate_cumulative_survival(survival_rates, initial_population = 1)
}
\arguments{
\item{survival_rates}{Vector de tasas de supervivencia diarias}

\item{initial_population}{Población inicial}
}
\value{
Vector con población superviviente por día
}
\description{
Calcula la supervivencia a lo largo de múltiples períodos
}
