% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{calculate_daily_growth}
\alias{calculate_daily_growth}
\title{Calculate final daily weight considering growth and reproduction}
\usage{
calculate_daily_growth(
  current_weight,
  net_energy,
  spawn_energy,
  predator_energy_density
)
}
\arguments{
\item{current_weight}{Initial weight of the day (g)}

\item{net_energy}{Net energy available (J/g/day)}

\item{spawn_energy}{Energy lost to reproduction (J)}

\item{predator_energy_density}{Predator energy density (J/g)}
}
\value{
List with final weight and weight change
}
\description{
Calculate final daily weight considering growth and reproduction
}
\keyword{internal}
