% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{calculate_daily_respiration}
\alias{calculate_daily_respiration}
\title{Calcular respiración diaria}
\usage{
calculate_daily_respiration(weight, temperature, consumption, species_params)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{temperature}{Temperatura (°C)}

\item{consumption}{Consumo del día}

\item{species_params}{Parámetros de la especie}
}
\value{
Respiración diaria
}
\description{
Calcular respiración diaria
}
\keyword{internal}
