% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{calculate_ecosystem_impact}
\alias{calculate_ecosystem_impact}
\title{Estimate ecosystem impact of nutrient excretion}
\usage{
calculate_ecosystem_impact(
  daily_n_excretion,
  daily_p_excretion,
  fish_biomass,
  water_volume,
  simulation_days
)
}
\arguments{
\item{daily_n_excretion}{Daily nitrogen excretion (g N/day)}

\item{daily_p_excretion}{Daily phosphorus excretion (g P/day)}

\item{fish_biomass}{Total fish biomass in system (g)}

\item{water_volume}{Water volume of system (L)}

\item{simulation_days}{Number of simulation days}
}
\value{
List with ecosystem impact estimates
}
\description{
Calculates potential ecosystem impact of nutrient excretion
}
