% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_efficiencies}
\alias{calculate_efficiencies}
\title{Calcular eficiencias de asimilación y retención}
\usage{
calculate_efficiencies(consumption, egestion, excretion)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{egestion}{Egestión (J/g)}

\item{excretion}{Excreción (J/g)}
}
\value{
Lista con eficiencias calculadas
}
\description{
Calcular eficiencias de asimilación y retención
}
