% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_egestion}
\alias{calculate_egestion}
\title{Calculate daily egestion (Mid-level - Main function)}
\usage{
calculate_egestion(
  consumption,
  temperature,
  p_value,
  processed_egestion_params
)
}
\arguments{
\item{consumption}{Consumption (J/g)}

\item{temperature}{Water temperature (deg C)}

\item{p_value}{Proportion of maximum consumption (p_value)}

\item{processed_egestion_params}{List with processed egestion parameters}
}
\value{
Egestion (J/g)
}
\description{
Main egestion calculation function called from simulation loop
}
