% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_egestion_rate}
\alias{calculate_egestion_rate}
\title{Calcular tasa de egestión}
\usage{
calculate_egestion_rate(consumption, temperature, p_value, egestion_params)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{temperature}{Temperatura (°C)}

\item{p_value}{P-value}

\item{egestion_params}{Parámetros de egestión}
}
\value{
Tasa de egestión (J/g)
}
\description{
Calcular tasa de egestión
}
\keyword{internal}
