% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{calculate_energy_balance}
\alias{calculate_energy_balance}
\title{Calcular balance energético}
\usage{
calculate_energy_balance(
  initial_weight,
  final_weight,
  total_spawn_energy,
  daily_output = NULL,
  predator_energy_params
)
}
\arguments{
\item{initial_weight}{Peso inicial}

\item{final_weight}{Peso final}

\item{total_spawn_energy}{Energía total perdida por reproducción}

\item{daily_output}{Salida diaria (opcional)}

\item{predator_energy_params}{Parámetros de energía del depredador}
}
\value{
Lista con información del balance energético
}
\description{
Calcular balance energético
}
\keyword{internal}
