% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{calculate_energy_density}
\alias{calculate_energy_density}
\title{Calculate energy density from fat and protein (Low-level)}
\usage{
calculate_energy_density(
  fat_content,
  protein_content,
  total_weight,
  fat_energy = 36200,
  protein_energy = 23600
)
}
\arguments{
\item{fat_content}{Fat content (g)}

\item{protein_content}{Protein content (g)}

\item{total_weight}{Total weight (g)}

\item{fat_energy}{Energy per gram of fat (J/g)}

\item{protein_energy}{Energy per gram of protein (J/g)}
}
\value{
Energy density (J/g wet weight)
}
\description{
Calculates energy density based on fat and protein content
}
\keyword{internal}
