% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_excretion_rate}
\alias{calculate_excretion_rate}
\title{Calcular tasa de excreción}
\usage{
calculate_excretion_rate(
  consumption,
  egestion,
  temperature,
  p_value,
  excretion_params
)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{egestion}{Egestión (J/g)}

\item{temperature}{Temperatura (°C)}

\item{p_value}{P-value}

\item{excretion_params}{Parámetros de excreción}
}
\value{
Tasa de excreción (J/g)
}
\description{
Calcular tasa de excreción
}
\keyword{internal}
