% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{calculate_fat_by_subtraction}
\alias{calculate_fat_by_subtraction}
\title{Calculate fat content by subtraction (Low-level)}
\usage{
calculate_fat_by_subtraction(
  total_weight,
  water_content,
  protein_content,
  ash_content
)
}
\arguments{
\item{total_weight}{Total wet weight (g)}

\item{water_content}{Water content (g)}

\item{protein_content}{Protein content (g)}

\item{ash_content}{Ash content (g)}
}
\value{
Fat content (g)
}
\description{
Calculates grams of fat by subtracting water, protein and ash from total weight
}
\keyword{internal}
