% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{calculate_final_weight}
\alias{calculate_final_weight}
\title{Calcular peso final dados energía ganada y densidad energética}
\usage{
calculate_final_weight(
  initial_weight,
  energy_gain,
  predator_params,
  spawn_energy = 0
)
}
\arguments{
\item{initial_weight}{Peso inicial (g)}

\item{energy_gain}{Ganancia neta de energía (J)}

\item{predator_params}{Lista con parámetros de densidad energética}

\item{spawn_energy}{Energía perdida por reproducción (J)}
}
\value{
Lista con peso final y densidad energética final
}
\description{
Calcular peso final dados energía ganada y densidad energética
}
