% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{calculate_final_weight_with_energy_density}
\alias{calculate_final_weight_with_energy_density}
\title{Calcular peso final dados energía ganada y densidad energética}
\usage{
calculate_final_weight_with_energy_density(
  initial_weight,
  energy_gain,
  spawn_energy = 0,
  PREDEDEQ,
  alpha1 = NULL,
  beta1 = NULL,
  alpha2 = NULL,
  beta2 = NULL,
  cutoff = NULL,
  initial_energy_density = NULL,
  final_energy_density = NULL
)
}
\arguments{
\item{initial_weight}{Peso inicial (g)}

\item{energy_gain}{Ganancia neta de energía (J)}

\item{spawn_energy}{Energía perdida por reproducción (J)}

\item{PREDEDEQ}{Ecuación de densidad energética}

\item{alpha1}{Parámetro alpha1}

\item{beta1}{Parámetro beta1}

\item{alpha2}{Parámetro alpha2 (para PREDEDEQ = 2)}

\item{beta2}{Parámetro beta2 (para PREDEDEQ = 2)}

\item{cutoff}{Peso de corte (para PREDEDEQ = 2)}

\item{initial_energy_density}{Densidad energética inicial (para PREDEDEQ = 1)}

\item{final_energy_density}{Densidad energética final (para PREDEDEQ = 1)}
}
\value{
Lista con peso final y densidad energética final
}
\description{
Función auxiliar para calcular el peso final del pez usando diferentes
ecuaciones de densidad energética
}
