% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{calculate_fish_water_partition}
\alias{calculate_fish_water_partition}
\title{Calculate fish:water partition coefficient (Utility)}
\usage{
calculate_fish_water_partition(
  fat_fraction,
  protein_ash_fraction,
  water_fraction,
  kow
)
}
\arguments{
\item{fat_fraction}{Fat fraction in fish}

\item{protein_ash_fraction}{Protein + ash fraction}

\item{water_fraction}{Water fraction}

\item{kow}{Octanol:water partition coefficient}
}
\value{
Fish:water partition coefficient
}
\description{
Calculates Kbw based on body composition and Kow according to Arnot & Gobas (2004)
}
