% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{calculate_gill_efficiency}
\alias{calculate_gill_efficiency}
\title{Calculate gill uptake efficiency (Utility)}
\usage{
calculate_gill_efficiency(kow)
}
\arguments{
\item{kow}{Octanol:water partition coefficient}
}
\value{
Gill uptake efficiency
}
\description{
Calculates gill efficiency based on Kow according to Arnot & Gobas (2004)
}
