% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{calculate_maximum_consumption}
\alias{calculate_maximum_consumption}
\title{Calcular consumo máximo}
\usage{
calculate_maximum_consumption(weight, temperature, species_params)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{temperature}{Temperatura (°C)}

\item{species_params}{Parámetros de la especie}
}
\value{
Consumo máximo (g/g/día)
}
\description{
Calcular consumo máximo
}
\keyword{internal}
