% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R,
%   R/11-data-processing.R
\name{calculate_mortality_intervals}
\alias{calculate_mortality_intervals}
\title{Calcular intervalos de mortalidad}
\usage{
calculate_mortality_intervals(mortality_rates, mort_types)

calculate_mortality_intervals(mortality_rates, mort_types)
}
\arguments{
\item{mortality_rates}{Tasas de mortalidad}

\item{mort_types}{Tipos de mortalidad}

\item{mortality_types}{Nombres de tipos de mortalidad}
}
\value{
Matriz con intervalos de mortalidad

Array con intervalos de mortalidad
}
\description{
Determina la duración de períodos con tasas de mortalidad constantes
}
\keyword{internal}
