% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_net_retention_efficiency}
\alias{calculate_net_retention_efficiency}
\title{Calcular eficiencia de retenci贸n neta}
\usage{
calculate_net_retention_efficiency(C, Eg, Ex)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Eg}{Egestion (J/g)}

\item{Ex}{Excreci贸n (J/g)}
}
\value{
Eficiencia de retenci贸n neta (fracci贸n)
}
\description{
Calcula la eficiencia con la que la materia asimilada es retenida
}
