% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{calculate_nutrient_allocation}
\alias{calculate_nutrient_allocation}
\title{Generic nutrient allocation in bioenergetic model (Low-level)}
\usage{
calculate_nutrient_allocation(
  consumption,
  prey_nutrient_concentrations,
  nutrient_assimilation_efficiency,
  weight_gain,
  predator_nutrient_concentration,
  nutrient_name = "nutrient"
)
}
\arguments{
\item{consumption}{Vector of consumption by prey type (g/day)}

\item{prey_nutrient_concentrations}{Vector of nutrient concentrations in prey (g nutrient/g wet weight)}

\item{nutrient_assimilation_efficiency}{Vector of nutrient assimilation efficiencies (fraction 0-1)}

\item{weight_gain}{Predator weight gain (g/day)}

\item{predator_nutrient_concentration}{Nutrient concentration in predator (g nutrient/g wet weight)}

\item{nutrient_name}{Name of nutrient for error messages}
}
\value{
List with nutrient fluxes
}
\description{
Calculates nutrient balance in consumption, growth, excretion and egestion
}
\keyword{internal}
