% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R,
%   R/11-data-processing.R
\name{calculate_population_survival}
\alias{calculate_population_survival}
\title{Calcular supervivencia poblacional}
\usage{
calculate_population_survival(mortality_rates, mortality_intervals)

calculate_population_survival(mortality_rates, mortality_intervals)
}
\arguments{
\item{mortality_rates}{Tasas de mortalidad}

\item{mortality_intervals}{Intervalos de mortalidad}

\item{initial_population}{Población inicial}
}
\value{
Lista con dinámicas poblacionales

Vector con supervivencia poblacional
}
\description{
Calcula la dinámica poblacional considerando múltiples fuentes de mortalidad
}
\keyword{internal}
