% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-simulation-engine.R
\name{calculate_progress_interval}
\alias{calculate_progress_interval}
\title{Calculate appropriate progress reporting interval}
\usage{
calculate_progress_interval(n_days)
}
\description{
Calculate appropriate progress reporting interval
}
\keyword{internal}
