% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{calculate_protein_from_water}
\alias{calculate_protein_from_water}
\title{Estimate protein content from water (Low-level)}
\usage{
calculate_protein_from_water(water_content)
}
\arguments{
\item{water_content}{Water content (g)}
}
\value{
Protein content (g)
}
\description{
Estimates grams of protein based on water content using Breck (2014) regression
}
\references{
Breck, J.E. 2014. Body composition in fishes: body size matters. Aquaculture 433:40-49.
}
\keyword{internal}
