% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{calculate_reproductive_loss}
\alias{calculate_reproductive_loss}
\title{Calculate reproductive weight loss (Low-level)}
\usage{
calculate_reproductive_loss(
  spawn_fraction,
  current_weight,
  energy_density = 5000
)
}
\arguments{
\item{spawn_fraction}{Fraction of weight lost in reproduction (0-1)}

\item{current_weight}{Current fish weight (g)}

\item{energy_density}{Energy density of reproductive tissue (J/g)}
}
\value{
List with weight and energy losses
}
\description{
Calculates weight and energy loss during reproductive events
}
\keyword{internal}
