% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_respiration_params_eq2}
\alias{calculate_respiration_params_eq2}
\title{Calculate additional parameters for respiration equation 2 (Mid-level)}
\usage{
calculate_respiration_params_eq2(RQ, RTM, RTO)
}
\arguments{
\item{RQ}{Q10 rate}

\item{RTM}{Maximum temperature}

\item{RTO}{Optimum temperature}
}
\value{
List with RY, RZ, RX
}
\description{
Calculates derived parameters needed for temperature equation 2
}
