% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_respiration_rate}
\alias{calculate_respiration_rate}
\title{Calcular tasa de respiración}
\usage{
calculate_respiration_rate(weight, temperature, respiration_params)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{temperature}{Temperatura del agua (°C)}

\item{respiration_params}{Parámetros de respiración}
}
\value{
Tasa de respiración (g O2/g/día)
}
\description{
Calcular tasa de respiración
}
\keyword{internal}
