% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_sda}
\alias{calculate_sda}
\title{Calcular Acción Dinámica Específica (SDA)}
\usage{
calculate_sda(consumption_energy, egestion_energy, SDA_coeff)
}
\arguments{
\item{consumption_energy}{Consumo en energía (J/g)}

\item{egestion_energy}{Egestión en energía (J/g)}

\item{SDA_coeff}{Coeficiente de acción dinámica específica}
}
\value{
SDA en energía (J/g)
}
\description{
Calcular Acción Dinámica Específica (SDA)
}
