% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_simple_activity}
\alias{calculate_simple_activity}
\title{Función simplificada de actividad (constante)}
\usage{
calculate_simple_activity(ACT)
}
\arguments{
\item{ACT}{Multiplicador de actividad constante}
}
\value{
Factor de actividad
}
\description{
Función simplificada de actividad (constante)
}
\keyword{internal}
