% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{calculate_spawn_energy}
\alias{calculate_spawn_energy}
\title{Calculate spawning energy loss (Low-level)}
\usage{
calculate_spawn_energy(spawn_fraction, current_weight, energy_density)
}
\arguments{
\item{spawn_fraction}{Fraction of weight lost in reproduction (0-1)}

\item{current_weight}{Current fish weight (g)}

\item{energy_density}{Energy density of reproductive tissue (J/g)}
}
\value{
Spawning energy loss (J)
}
\description{
Calculates energy lost to reproduction on a given day
}
\keyword{internal}
