% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{calculate_temperature_factor_consumption}
\alias{calculate_temperature_factor_consumption}
\title{Calculate temperature factor for consumption (Mid-level)}
\usage{
calculate_temperature_factor_consumption(temperature, consumption_params)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{consumption_params}{List with consumption parameters}
}
\value{
Temperature factor
}
\description{
Coordinates temperature equation selection and calculation
}
\keyword{internal}
