% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{calibrate_nutrient_assimilation}
\alias{calibrate_nutrient_assimilation}
\title{Calibrar eficiencias de asimilación de nutrientes}
\usage{
calibrate_nutrient_assimilation(
  consumption,
  prey_concentrations,
  predator_concentration,
  weight_gain,
  observed_excretion,
  initial_efficiency,
  nutrient = "nitrogen"
)
}
\arguments{
\item{consumption}{Vector de consumo por presa}

\item{prey_concentrations}{Concentraciones en presas}

\item{predator_concentration}{Concentración en depredador}

\item{weight_gain}{Ganancia de peso}

\item{observed_excretion}{Excreción observada}

\item{initial_efficiency}{Eficiencias iniciales}

\item{nutrient}{Tipo de nutriente ("nitrogen" o "phosphorus")}
}
\value{
Lista con eficiencias calibradas
}
\description{
Ajusta eficiencias de asimilación para que coincidan con excreción observada
}
