% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{Cf1T}
\alias{Cf1T}
\title{Funci처n de temperatura para consumo - Ecuaci처n 1}
\usage{
Cf1T(Temperature, CQ)
}
\arguments{
\item{Temperature}{Temperatura del agua (째C)}

\item{CQ}{Coeficiente de temperatura para consumo}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuaci처n de temperatura 1 (Hanson et al. 1997; Stewart et al. 1983)
}
