% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{Cf2T}
\alias{Cf2T}
\title{Función de temperatura para consumo - Ecuación 2}
\usage{
Cf2T(Temperature, CTM, CTO, CX)
}
\arguments{
\item{Temperature}{Temperatura del agua (°C)}

\item{CTM}{Temperatura máxima donde cesa el consumo}

\item{CTO}{Temperatura óptima de laboratorio}

\item{CX}{Parámetro calculado}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuación de temperatura 2 (Hanson et al. 1997; Kitchell et al. 1977)
}
