% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{Cf4T}
\alias{Cf4T}
\title{Funci처n de temperatura para consumo - Ecuaci처n 4}
\usage{
Cf4T(Temperature, CQ, CK1, CK4)
}
\arguments{
\item{Temperature}{Temperatura del agua (째C)}

\item{CQ}{Coeficiente de temperatura}

\item{CK1}{Coeficiente de segundo orden}

\item{CK4}{Coeficiente de tercer orden}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuaci처n de temperatura 4 (Bevelhimer et al. 1985)
}
