% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{check_egestion_excretion_balance}
\alias{check_egestion_excretion_balance}
\title{Calcular balance energético para egestion y excreción}
\usage{
check_egestion_excretion_balance(C, Eg, Ex, tolerance = 0.01)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Eg}{Egestion (J/g)}

\item{Ex}{Excreción (J/g)}

\item{tolerance}{Tolerancia para el balance (fracción)}
}
\value{
Lista con información del balance
}
\description{
Verifica que las pérdidas no excedan el consumo
}
