% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.0-tmb-shared.R
\name{check_tmb_compilation}
\alias{check_tmb_compilation}
\title{Check TMB model availability (simplified for installed package)}
\usage{
check_tmb_compilation(dll_name = "fb4package", verbose = FALSE)
}
\arguments{
\item{dll_name}{Name of DLL (default: "fb4package")}

\item{verbose}{Show diagnostic information}
}
\value{
TRUE if model is available, FALSE otherwise
}
\description{
Verify that TMB model is available in installed package.
}
\keyword{internal}
