% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.3-data-validators.R
\name{check_weight_outliers}
\alias{check_weight_outliers}
\title{Check for outliers in weight data}
\usage{
check_weight_outliers(individual_data)
}
\arguments{
\item{individual_data}{Individual data frame}
}
\description{
Identifies potential outliers in weight measurements using IQR method.
}
\keyword{internal}
