% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{classify_spawn_season}
\alias{classify_spawn_season}
\title{Clasificar temporada de desove}
\usage{
classify_spawn_season(spawn_centroid)
}
\arguments{
\item{spawn_centroid}{Día centroide del desove}
}
\value{
Temporada reproductiva
}
\description{
Clasifica la temporada reproductiva basada en el timing
}
\keyword{internal}
