% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.1-plot-core.R
\name{close_save_device}
\alias{close_save_device}
\title{Close graphics device after saving}
\usage{
close_save_device(save_path)
}
\arguments{
\item{save_path}{Path where file was saved}
}
\description{
Closes graphics device and shows confirmation message.
}
\keyword{internal}
