% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{compare_body_compositions}
\alias{compare_body_compositions}
\title{Comparar composici贸n corporal entre especies/condiciones}
\usage{
compare_body_compositions(conditions, weight = 100)
}
\arguments{
\item{conditions}{Lista con condiciones a comparar}

\item{weight}{Peso de referencia para comparaci贸n}
}
\value{
Data frame con comparaci贸n de composiciones
}
\description{
Compara composici贸n corporal bajo diferentes condiciones
}
