% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{consumption_temp_eq1}
\alias{consumption_temp_eq1}
\title{Temperature function for consumption - Equation 1 (Low-level)}
\usage{
consumption_temp_eq1(temperature, CQ)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{CQ}{Temperature coefficient for consumption}
}
\value{
Temperature factor
}
\description{
Implements temperature equation 1 (simple exponential)
}
\keyword{internal}
