% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{consumption_temp_eq3}
\alias{consumption_temp_eq3}
\title{Función de temperatura para consumo - Ecuación 3}
\usage{
consumption_temp_eq3(temperature, CQ, CG1, CK1, CG2, CTL, CK4)
}
\arguments{
\item{temperature}{Temperatura del agua (°C)}

\item{CQ}{Coeficiente de temperatura}

\item{CG1}{Parámetro calculado 1}

\item{CK1}{Fracción pequeña de la tasa máxima}

\item{CG2}{Parámetro calculado 2}

\item{CTL}{Temperatura con tasa reducida}

\item{CK4}{Fracción reducida de la tasa máxima}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuación de temperatura 3 (Thornton y Lessem 1978)
}
\keyword{internal}
