% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{consumption_temp_eq4}
\alias{consumption_temp_eq4}
\title{Función de temperatura para consumo - Ecuación 4}
\usage{
consumption_temp_eq4(temperature, CQ, CK1, CK4)
}
\arguments{
\item{temperature}{Temperatura del agua (°C)}

\item{CQ}{Coeficiente lineal}

\item{CK1}{Coeficiente cuadrático}

\item{CK4}{Coeficiente cúbico}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuación de temperatura 4 (polinomial)
}
\keyword{internal}
