% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{contaminant_model_1}
\alias{contaminant_model_1}
\title{Modelo de contaminantes 1 - Solo captación de alimento}
\usage{
contaminant_model_1(C, X_Prey, TEx, Burden)
}
\arguments{
\item{C}{Vector de consumo por presa (g/día)}

\item{X_Prey}{Vector de concentraciones en presas (μg/g)}

\item{TEx}{Vector de eficiencias de transferencia}

\item{Burden}{Carga corporal actual (μg)}
}
\value{
Vector con c(Clearance, Uptake, Burden, NA)
}
\description{
Modelo simple sin eliminación, solo acumulación desde el alimento
}
