% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{contaminant_model_2}
\alias{contaminant_model_2}
\title{Contaminant model 2 - With temperature and weight dependent elimination (Low-level)}
\usage{
contaminant_model_2(
  consumption,
  weight,
  temperature,
  prey_concentrations,
  assimilation_efficiency,
  current_burden
)
}
\arguments{
\item{consumption}{Vector of consumption by prey (g/day)}

\item{weight}{Fish weight (g)}

\item{temperature}{Water temperature (deg C)}

\item{prey_concentrations}{Vector of concentrations in prey (ug/g)}

\item{assimilation_efficiency}{Vector of assimilation efficiencies}

\item{current_burden}{Current body burden (ug)}
}
\value{
List with clearance, uptake, and new burden
}
\description{
Model with food uptake and elimination dependent on temperature and weight
Based on Trudel & Rasmussen (1997) for MeHg
}
\keyword{internal}
