% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{contaminant_model_3}
\alias{contaminant_model_3}
\title{Modelo de contaminantes 3 - Arnot & Gobas (2004)}
\usage{
contaminant_model_3(
  R_O2,
  C,
  W,
  Temperature,
  X_Prey,
  X_ae,
  Burden,
  Ew,
  Kbw,
  Cw_tot,
  phi_DT,
  DO_Sat_fr
)
}
\arguments{
\item{R_O2}{Respiración (g O2/g/día)}

\item{C}{Vector de consumo por presa (g/día)}

\item{W}{Peso del pez (g)}

\item{Temperature}{Temperatura del agua (°C)}

\item{X_Prey}{Vector de concentraciones en presas (μg/g)}

\item{X_ae}{Vector de eficiencias de asimilación}

\item{Burden}{Carga corporal actual (μg)}

\item{Ew}{Eficiencia de captación branquial}

\item{Kbw}{Coeficiente de partición pez:agua}

\item{Cw_tot}{Concentración total en agua (mg/L)}

\item{phi_DT}{Fracción disuelta}

\item{DO_Sat_fr}{Saturación de oxígeno (fracción)}
}
\value{
Vector con c(Clearance, Uptake, Burden, NA)
}
\description{
Modelo completo con captación desde agua y alimento, eliminación proporcional
a la respiración
}
