% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-simulation-engine.R
\name{convert_daily_results_to_dataframe}
\alias{convert_daily_results_to_dataframe}
\title{Convert list of daily results to data frame}
\usage{
convert_daily_results_to_dataframe(daily_results)
}
\description{
Convert list of daily results to data frame
}
\keyword{internal}
