% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{convert_respiration_to_energy}
\alias{convert_respiration_to_energy}
\title{Convert respiration from O2 to energy units (Utility)}
\usage{
convert_respiration_to_energy(respiration_o2, oxycal = 13560)
}
\arguments{
\item{respiration_o2}{Respiration in g O2/g fish/day}

\item{oxycal}{Oxycalorific conversion factor (J/g O2)}
}
\value{
Respiration in J/g fish/day
}
\description{
Converts respiration from O2 consumption to energy equivalents
}
