% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.2.2-strategy-binary-search.R
\name{create_binary_search_strategy}
\alias{create_binary_search_strategy}
\title{Create Binary Search Strategy}
\usage{
create_binary_search_strategy(execution_plan)
}
\arguments{
\item{execution_plan}{Execution plan with parameters}
}
\value{
Strategy object implementing FB4Strategy interface
}
\description{
Create Binary Search Strategy
}
\keyword{internal}
