% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.2.4-strategy-bootstrap.R
\name{create_bootstrap_strategy}
\alias{create_bootstrap_strategy}
\title{Create Bootstrap Strategy}
\usage{
create_bootstrap_strategy(execution_plan)
}
\arguments{
\item{execution_plan}{Execution plan with parameters}
}
\value{
Strategy object implementing FB4Strategy interface
}
\description{
Create Bootstrap Strategy
}
\keyword{internal}
