% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.2.1-strategy-direct.R
\name{create_direct_strategy}
\alias{create_direct_strategy}
\title{Create Direct Strategy}
\usage{
create_direct_strategy(execution_plan, direct_type)
}
\arguments{
\item{execution_plan}{Execution plan with parameters}

\item{direct_type}{Type of direct method ("p_value", "ration_percent", "ration_grams")}
}
\value{
Strategy object implementing FB4Strategy interface
}
\description{
Create Direct Strategy
}
\keyword{internal}
