% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.1-basic-validators.R
\name{create_empty_composition}
\alias{create_empty_composition}
\title{Create empty composition for invalid inputs (Utility)}
\usage{
create_empty_composition()
}
\value{
Empty composition list
}
\description{
Create empty composition for invalid inputs (Utility)
}
\keyword{internal}
