% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{create_failed_result}
\alias{create_failed_result}
\title{Crear resultado fallido mínimo}
\usage{
create_failed_result(initial_weight, n_days)
}
\arguments{
\item{initial_weight}{Peso inicial}

\item{n_days}{Número de días}
}
\value{
Resultado mínimo para casos de falla
}
\description{
Crear resultado fallido mínimo
}
\keyword{internal}
