% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{create_fb4_dashboard}
\alias{create_fb4_dashboard}
\title{Crear dashboard de resultados}
\usage{
create_fb4_dashboard(fb4_result, save_path = NULL)
}
\arguments{
\item{fb4_result}{Objeto fb4_result}

\item{save_path}{Ruta para guardar el gráfico (opcional)}
}
\description{
Crear dashboard de resultados
}
