% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.2.6-strategy-hierarchical.R
\name{create_hierarchical_strategy}
\alias{create_hierarchical_strategy}
\title{Create Hierarchical Strategy}
\usage{
create_hierarchical_strategy(execution_plan)
}
\arguments{
\item{execution_plan}{Execution plan with parameters}
}
\value{
Strategy object implementing FB4Strategy interface
}
\description{
Create Hierarchical Strategy
}
\keyword{internal}
