% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.2.5-strategy-mle.R
\name{create_mle_strategy}
\alias{create_mle_strategy}
\title{Create MLE Strategy}
\usage{
create_mle_strategy(execution_plan)
}
\arguments{
\item{execution_plan}{Execution plan with parameters}
}
\value{
Strategy object implementing FB4Strategy interface
}
\description{
Create MLE Strategy
}
\keyword{internal}
