% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.2.3-strategy-optim.R
\name{create_optim_strategy}
\alias{create_optim_strategy}
\title{Create Optim Strategy}
\usage{
create_optim_strategy(execution_plan)
}
\arguments{
\item{execution_plan}{Execution plan with parameters}
}
\value{
Strategy object implementing FB4Strategy interface
}
\description{
Create Optim Strategy
}
\keyword{internal}
