% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.3-result-builders.R
\name{create_unified_fit_info}
\alias{create_unified_fit_info}
\title{Create unified fit info section}
\usage{
create_unified_fit_info(raw_results, execution_plan, method)
}
\description{
Create unified fit info section
}
\keyword{internal}
