% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion1}
\alias{egestion1}
\title{Modelo de egestion 1 - Básico}
\usage{
egestion1(C, FA)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{FA}{Fracción de egestion}
}
\value{
Egestion (J/g)
}
\description{
Modelo simple de egestion con fracción constante del consumo
}
