% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion2}
\alias{egestion2}
\title{Modelo de egestion 2 - Elliott (1976)}
\usage{
egestion2(C, Temperature, p, FA, FB, FG)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Temperature}{Temperatura del agua (°C)}

\item{p}{Proporción del consumo máximo (p-value)}

\item{FA}{Parámetro base de egestion}

\item{FB}{Coeficiente de dependencia de temperatura}

\item{FG}{Coeficiente de dependencia del nivel de alimentación}
}
\value{
Egestion (J/g)
}
\description{
Modelo de egestion dependiente de temperatura y nivel de alimentación
}
