% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion3}
\alias{egestion3}
\title{Modelo de egestion 3 - Stewart et al. (1983)}
\usage{
egestion3(C, Temperature, p, FA, FB, FG, indigestible_fraction = 0)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Temperature}{Temperatura del agua (°C)}

\item{p}{Proporción del consumo máximo (p-value)}

\item{FA}{Parámetro base de egestion}

\item{FB}{Coeficiente de dependencia de temperatura}

\item{FG}{Coeficiente de dependencia del nivel de alimentación}

\item{indigestible_fraction}{Fracción indigerible de las presas}
}
\value{
Egestion (J/g)
}
\description{
Modelo que incluye presas indigeribles
}
