% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion4}
\alias{egestion4}
\title{Modelo de egestion 4 - Elliott (1976) sin p-value}
\usage{
egestion4(C, Temperature, FA, FB)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Temperature}{Temperatura del agua (°C)}

\item{FA}{Parámetro base de egestion}

\item{FB}{Coeficiente de dependencia de temperatura}
}
\value{
Egestion (J/g)
}
\description{
Modelo simplificado sin dependencia del nivel de alimentación
}
