% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion_model_1}
\alias{egestion_model_1}
\title{Modelo de egestión 1 - Básico}
\usage{
egestion_model_1(consumption, FA)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{FA}{Fracción de egestión}
}
\value{
Egestión (J/g)
}
\description{
Modelo simple de egestión con fracción constante del consumo
}
\keyword{internal}
