% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion_model_3}
\alias{egestion_model_3}
\title{Egestion model 3 - Stewart et al. (1983) (Low-level)}
\usage{
egestion_model_3(
  consumption,
  temperature,
  p_value,
  FA,
  FB,
  FG,
  indigestible_fraction = 0
)
}
\arguments{
\item{consumption}{Consumption (J/g)}

\item{temperature}{Water temperature (°C)}

\item{p_value}{Proportion of maximum consumption (p-value)}

\item{FA}{Base egestion parameter}

\item{FB}{Temperature dependence coefficient}

\item{FG}{Feeding level dependence coefficient}

\item{indigestible_fraction}{Indigestible fraction of prey}
}
\value{
Egestion (J/g)
}
\description{
Model that includes indigestible prey
}
\keyword{internal}
