% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{EnDen}
\alias{EnDen}
\title{Estimar densidad energética a partir de grasa y proteína}
\usage{
EnDen(Fat_g, Pro_g, W, fat_energy = 39300, protein_energy = 23600)
}
\arguments{
\item{Fat_g}{Contenido de grasa (g)}

\item{Pro_g}{Contenido de proteína (g)}

\item{W}{Peso total (g)}

\item{fat_energy}{Energía por gramo de grasa (J/g), por defecto 39300}

\item{protein_energy}{Energía por gramo de proteína (J/g), por defecto 23600}
}
\value{
Densidad energética (J/g peso húmedo)
}
\description{
Calcula densidad energética basada en contenido de grasa y proteína
}
\examples{
W <- 100
H2O <- 72
Pro <- Protein(H2O)
Ash_content <- Ash(H2O)
Fat_content <- Fat(W, H2O, Pro, Ash_content)
ED <- EnDen(Fat_content, Pro, W)
print(ED)
}
