% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\docType{data}
\name{EQUATION_REQUIREMENTS}
\alias{EQUATION_REQUIREMENTS}
\title{Equation requirements for all FB4 components}
\format{
An object of class \code{list} of length 5.
}
\usage{
EQUATION_REQUIREMENTS
}
\description{
Equation requirements for all FB4 components
}
\keyword{internal}
