% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{estimate_energy_density_simple}
\alias{estimate_energy_density_simple}
\title{Estimar densidad energética simplificada}
\usage{
estimate_energy_density_simple(weight, water_fraction = 0.728)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{water_fraction}{Fracción de agua}
}
\value{
Densidad energética (J/g)
}
\description{
Función simplificada para estimar densidad energética cuando no se necesita
composición corporal completa
}
