% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{estimate_lipid_fraction}
\alias{estimate_lipid_fraction}
\title{Estimar fracción lipídica para modelos de contaminantes}
\usage{
estimate_lipid_fraction(weight, water_fraction = 0.728)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{water_fraction}{Fracción de agua}
}
\value{
Fracción lipídica
}
\description{
Estimar fracción lipídica para modelos de contaminantes
}
