% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{excretion1}
\alias{excretion1}
\title{Modelo de excreción 1 - Básico}
\usage{
excretion1(C, Eg, UA)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Eg}{Egestion (J/g)}

\item{UA}{Fracción de excreción}
}
\value{
Excreción (J/g)
}
\description{
Modelo simple de excreción proporcional a la materia asimilada
}
