% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{excretion3}
\alias{excretion3}
\title{Modelo de excreción 3 - Variante del modelo 2}
\usage{
excretion3(C, Temperature, p, Eg, UA, UB, UG)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Temperature}{Temperatura del agua (°C)}

\item{p}{Proporción del consumo máximo (p-value)}

\item{Eg}{Egestion (J/g)}

\item{UA}{Parámetro base de excreción}

\item{UB}{Coeficiente de dependencia de temperatura}

\item{UG}{Coeficiente de dependencia del nivel de alimentación}
}
\value{
Excreción (J/g)
}
\description{
Similar al modelo 2 pero con formulación ligeramente diferente
}
