% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{excretion4}
\alias{excretion4}
\title{Modelo de excreción 4 - Sin dependencia de alimentación}
\usage{
excretion4(C, Temperature, Eg, UA, UB)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Temperature}{Temperatura del agua (°C)}

\item{Eg}{Egestion (J/g)}

\item{UA}{Parámetro base de excreción}

\item{UB}{Coeficiente de dependencia de temperatura}
}
\value{
Excreción (J/g)
}
\description{
Modelo que solo depende de temperatura
}
