% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{excretion_model_1}
\alias{excretion_model_1}
\title{Excretion model 1 - Basic (Low-level)}
\usage{
excretion_model_1(consumption, egestion, UA)
}
\arguments{
\item{consumption}{Consumption (J/g)}

\item{egestion}{Egestion (J/g)}

\item{UA}{Excretion fraction}
}
\value{
Excretion (J/g)
}
\description{
Simple excretion model proportional to assimilated matter
}
\keyword{internal}
