% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{excretion_model_2}
\alias{excretion_model_2}
\title{Modelo de excreción 2 - Con dependencia de temperatura y alimentación}
\usage{
excretion_model_2(consumption, egestion, temperature, p_value, UA, UB, UG)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{egestion}{Egestión (J/g)}

\item{temperature}{Temperatura del agua (°C)}

\item{p_value}{Proporción del consumo máximo (p-value)}

\item{UA}{Parámetro base de excreción}

\item{UB}{Coeficiente de dependencia de temperatura}

\item{UG}{Coeficiente de dependencia del nivel de alimentación}
}
\value{
Excreción (J/g)
}
\description{
Modelo de excreción 2 - Con dependencia de temperatura y alimentación
}
\keyword{internal}
