% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{excretion_model_3}
\alias{excretion_model_3}
\title{Excretion model 3 - Variant of model 2 (Low-level)}
\usage{
excretion_model_3(consumption, egestion, temperature, p_value, UA, UB, UG)
}
\arguments{
\item{consumption}{Consumption (J/g)}

\item{egestion}{Egestion (J/g)}

\item{temperature}{Water temperature (°C)}

\item{p_value}{Proportion of maximum consumption (p-value)}

\item{UA}{Base excretion parameter}

\item{UB}{Temperature dependence coefficient}

\item{UG}{Feeding level dependence coefficient}
}
\value{
Excretion (J/g)
}
\description{
Alternative implementation of temperature and feeding dependent model
}
\keyword{internal}
