% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.2.5-strategy-mle.R
\name{execute_mle_tmb}
\alias{execute_mle_tmb}
\title{Execute MLE TMB using unified architecture}
\usage{
execute_mle_tmb(plan)
}
\description{
Execute MLE TMB using unified architecture
}
\keyword{internal}
