% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{export_fb4_results}
\alias{export_fb4_results}
\title{Exportar resultados a CSV}
\usage{
export_fb4_results(fb4_result, file_path, include_summary = TRUE)
}
\arguments{
\item{fb4_result}{Objeto fb4_result}

\item{file_path}{Ruta del archivo CSV}

\item{include_summary}{Incluir resumen en archivo separado}
}
\description{
Exportar resultados a CSV
}
