% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.0-tmb-shared.R
\name{extract_adreport_values}
\alias{extract_adreport_values}
\title{Extract ADREPORT values from TMB report}
\usage{
extract_adreport_values(report, model_type)
}
\arguments{
\item{report}{TMB report object}

\item{model_type}{Model type}
}
\value{
List of ADREPORT values
}
\description{
Extract ADREPORT values from TMB report
}
\keyword{internal}
