% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.0-tmb-shared.R
\name{extract_basic_parameters}
\alias{extract_basic_parameters}
\title{Extract parameters for basic model}
\usage{
extract_basic_parameters(results, params, obj, confidence_level, verbose)
}
\arguments{
\item{results}{Current results list}

\item{params}{Optimized parameters (log scale)}

\item{obj}{TMB objective function}

\item{confidence_level}{Confidence level}

\item{verbose}{Show progress}
}
\value{
Updated results list
}
\description{
Extract parameters for basic model
}
\keyword{internal}
