% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{extract_contaminant_day_data}
\alias{extract_contaminant_day_data}
\title{Extraer datos de contaminantes para un día específico}
\usage{
extract_contaminant_day_data(contaminant_data, day)
}
\arguments{
\item{contaminant_data}{Datos de contaminantes}

\item{day}{Día a extraer}
}
\value{
Lista con datos de contaminantes del día
}
\description{
Extraer datos de contaminantes para un día específico
}
\keyword{internal}
