% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-data-processing.R
\name{extract_data_sources}
\alias{extract_data_sources}
\title{Extract data sources from bioenergetic object}
\usage{
extract_data_sources(bio_obj)
}
\arguments{
\item{bio_obj}{Bioenergetic object}
}
\value{
List with data source information
}
\description{
Extract data sources from bioenergetic object
}
\keyword{internal}
