% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{extract_nutrient_day_data}
\alias{extract_nutrient_day_data}
\title{Extraer datos de nutrientes para un día específico}
\usage{
extract_nutrient_day_data(nutrient_data, day)
}
\arguments{
\item{nutrient_data}{Datos de nutrientes}

\item{day}{Día a extraer}
}
\value{
Lista con datos de nutrientes del día
}
\description{
Extraer datos de nutrientes para un día específico
}
\keyword{internal}
