% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-fb4.R, R/16.1-plot-core.R
\name{extract_species_name}
\alias{extract_species_name}
\title{Extract species name from fb4_result or Bioenergetic object}
\usage{
extract_species_name(fb4_result)

extract_species_name(fb4_result)
}
\arguments{
\item{fb4_result}{FB4 result object}
}
\value{
Character string with species name or NULL if not available
}
\description{
Extracts species name from fb4_result object for plot titles.
}
\keyword{internal}
