% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{extract_species_parameters}
\alias{extract_species_parameters}
\title{Extraer parámetros de especie}
\usage{
extract_species_parameters(parms, species_num)
}
\arguments{
\item{parms}{Data frame con parámetros de todas las especies}

\item{species_num}{Número de especie}
}
\value{
Lista con parámetros de la especie
}
\description{
Extraer parámetros de especie
}
\keyword{internal}
