% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{Fat}
\alias{Fat}
\title{Estimar contenido de grasa por sustracción}
\usage{
Fat(W, H2O, Pro, Ash)
}
\arguments{
\item{W}{Peso total húmedo (g)}

\item{H2O}{Contenido de agua (g)}

\item{Pro}{Contenido de proteína (g)}

\item{Ash}{Contenido de ceniza (g)}
}
\value{
Contenido de grasa (g)
}
\description{
Calcula gramos de grasa restando agua, proteína y ceniza del peso total
}
\examples{
W <- 100
H2O <- 72
Pro <- Protein(H2O)
Ash_content <- Ash(H2O)
Fat_content <- Fat(W, H2O, Pro, Ash_content)
print(Fat_content)
}
