% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb4package-package.R
\docType{package}
\name{fb4package-package}
\alias{fb4package}
\alias{fb4package-package}
\title{Fish Bioenergetics 4.0 Package}
\description{
An R implementation of fish bioenergetics modeling with TMB backend
for high-performance maximum likelihood estimation.
}
\author{
\strong{Maintainer}: Hans Ttito \email{kvttitos@gmail.com}

}
\keyword{internal}
