% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{fit}
\alias{fit}
\title{Ajustar modelo bioenergético}
\usage{
fit(x, target_value = NULL, fit_to = NULL, ...)
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{target_value}{Valor objetivo para el ajuste}

\item{fit_to}{Tipo de ajuste ("Weight", "Consumption", etc.)}

\item{...}{Argumentos adicionales}
}
\value{
Objeto Bioenergetic con resultados
}
\description{
Ajustar modelo bioenergético
}
