% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{fitting-algorithms-fb4}
\alias{fitting-algorithms-fb4}
\title{Algoritmos de Ajuste para el Modelo FB4 - Versión Integrada}
\description{
Algoritmos de Ajuste para el Modelo FB4 - Versión Integrada
}
