% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.1-plot-core.R
\name{get_available_plot_types}
\alias{get_available_plot_types}
\title{Get Available Plot Types}
\usage{
get_available_plot_types(x)
}
\arguments{
\item{x}{fb4_result object}
}
\value{
Character vector of available plot types
}
\description{
Returns list of available plot types for fb4_result objects.
}
\keyword{internal}
