% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{get_results}
\alias{get_results}
\title{Extraer resultados de un objeto Bioenergetic}
\usage{
get_results(x, component = "all")
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{component}{Componente a extraer ("daily_output", "summary", "all")}
}
\value{
Resultados extraídos
}
\description{
Extraer resultados de un objeto Bioenergetic
}
