% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.1-plot-core.R
\name{\%||\%}
\alias{\%||\%}
\title{Safe NULL coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value to check}

\item{y}{Fallback value if x is NULL}
}
\value{
First non-NULL value
}
\description{
Returns the first non-NULL value from a list of arguments.
}
\keyword{internal}
