% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.1-plot-core.R
\name{has_uncertainty}
\alias{has_uncertainty}
\title{Check if result has uncertainty information}
\usage{
has_uncertainty(x)
}
\arguments{
\item{x}{fb4_result object}
}
\value{
Logical indicating if uncertainty data is available
}
\description{
Checks if fb4_result contains uncertainty information.
}
\keyword{internal}
