% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{initialize_daily_output}
\alias{initialize_daily_output}
\title{Inicializar estructura de salida diaria}
\usage{
initialize_daily_output(n_days, prey_names)
}
\arguments{
\item{n_days}{Número de días}

\item{prey_names}{Nombres de las presas}
}
\value{
Data frame inicializado para salida diaria
}
\description{
Inicializar estructura de salida diaria
}
\keyword{internal}
