% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{instantaneous_to_mortality_rate}
\alias{instantaneous_to_mortality_rate}
\title{Convertir tasa instantánea a tasa de mortalidad}
\usage{
instantaneous_to_mortality_rate(Z, time_period = 1)
}
\arguments{
\item{Z}{Tasa de mortalidad instantánea}

\item{time_period}{Período de tiempo}
}
\value{
Tasa de mortalidad (fracción)
}
\description{
Convierte tasa instantánea (Z) a fracción que muere
}
\examples{
# Convertir Z = 0.5 año^-1 a mortalidad anual
annual_mortality <- instantaneous_to_mortality_rate(0.5, 365)
}
