% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{interpolate_missing_days}
\alias{interpolate_missing_days}
\title{Interpolar datos faltantes por días}
\usage{
interpolate_missing_days(data, all_days, value_cols, method = "linear")
}
\arguments{
\item{data}{Data frame con datos}

\item{all_days}{Vector con todos los días necesarios}

\item{value_cols}{Columnas a interpolar}

\item{method}{Método de interpolación: "linear" o "spline"}
}
\value{
Data frame con datos completos
}
\description{
Interpolar datos faltantes por días
}
\keyword{internal}
