% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0-bioenergetic-classes.R
\name{is.Bioenergetic}
\alias{is.Bioenergetic}
\title{Test if Object is Bioenergetic}
\usage{
is.Bioenergetic(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
A length-1 logical: \code{TRUE} if \code{x} inherits from class
  \code{"Bioenergetic"}, \code{FALSE} otherwise.
}
\description{
Tests whether an object inherits from the Bioenergetic class.
}
\examples{
bio <- Bioenergetic(
  species_params = list(
    consumption = list(CEQ = 1, CA = 0.303, CB = -0.275, CQ = 0.06)
  ),
  species_info = list(common_name = "Example fish")
)
is.Bioenergetic(bio)
is.Bioenergetic(list())
}
