% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.0-analysis-core.R
\name{is.fb4_result}
\alias{is.fb4_result}
\title{Test if Object is fb4_result}
\usage{
is.fb4_result(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
A length-1 logical: \code{TRUE} if \code{x} inherits from class
  \code{"fb4_result"}, \code{FALSE} otherwise.
}
\description{
Tests whether an object inherits from the fb4_result class.
}
\examples{
is.fb4_result(list())
}
