% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{list_species}
\alias{list_species}
\title{Listar especies disponibles en la base de datos}
\usage{
list_species(fish4_db = NULL, family = NULL)
}
\arguments{
\item{fish4_db}{Base de datos fish4_parameters (opcional)}

\item{family}{Filtrar por familia (opcional)}
}
\value{
Vector con nombres de especies
}
\description{
Listar especies disponibles en la base de datos
}
