% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{null-coalescing}
\alias{null-coalescing}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, y otherwise
}
\description{
Returns the first value if it's not NULL, otherwise returns the second value.
Similar to the || operator in other programming languages.
}
\details{
This operator is useful for providing default values when working with
potentially NULL parameters or list elements.
}
\examples{
# Basic usage
NULL \%||\% "default"        # Returns "default"
"value" \%||\% "default"     # Returns "value"

# Common use case with list elements
params <- list(temperature = 15)
temp <- params$temperature \%||\% 10  # Returns 15
depth <- params$depth \%||\% 5        # Returns 5 (default)
}
