% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to test}

\item{y}{Fallback value if \code{x} is \code{NULL}}
}
\value{
\code{x} if not \code{NULL}, otherwise \code{y}
}
\description{
Returns \code{x} if it is not \code{NULL}, otherwise returns \code{y}.
Equivalent to \code{rlang::\%||\%} but without requiring that dependency.
}
\examples{
NULL \%||\% "default"
"value" \%||\% "default"
list()$missing \%||\% 0
}
\keyword{internal}
