% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{optimize_fb4_parameters}
\alias{optimize_fb4_parameters}
\title{Optimizar parámetros del modelo usando múltiples objetivos}
\usage{
optimize_fb4_parameters(
  objectives,
  parameters_to_fit,
  initial_guess,
  species_params,
  processed_data,
  model_options,
  method = "Nelder-Mead",
  control = list(maxit = 100)
)
}
\arguments{
\item{objectives}{Lista de objetivos (ej: list(final_weight = 50, total_consumption = 100))}

\item{parameters_to_fit}{Vector con nombres de parámetros a ajustar}

\item{initial_guess}{Valores iniciales para parámetros}

\item{species_params}{Parámetros de la especie}

\item{processed_data}{Datos procesados}

\item{model_options}{Opciones del modelo}

\item{method}{Método de optimización ("Nelder-Mead", "BFGS", etc.)}

\item{control}{Lista de control para optimización}
}
\value{
Resultado de la optimización
}
\description{
Permite ajustar múltiples parámetros del modelo para cumplir varios objetivos
}
