% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.3-analysis-sensitivity.R
\name{perform_scenario_tests}
\alias{perform_scenario_tests}
\title{Perform statistical tests between scenarios}
\usage{
perform_scenario_tests(scenario_data, metrics)
}
\arguments{
\item{scenario_data}{Data frame with scenario results}

\item{metrics}{Vector of metrics to test}
}
\value{
List with test results
}
\description{
Performs statistical comparisons between scenarios when uncertainty
estimates are available.
}
\keyword{internal}
