% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-fb4.R
\name{plot.Bioenergetic}
\alias{plot.Bioenergetic}
\title{Plot Bioenergetic object setup}
\usage{
\method{plot}{Bioenergetic}(x, type = "dashboard", save_plot = NULL, ...)
}
\arguments{
\item{x}{Object of class Bioenergetic}

\item{type}{Type of plot: "dashboard", "temperature", "diet", "energy"}

\item{save_plot}{Optional path to save plot}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the input object
}
\description{
Plotting method for Bioenergetic objects to validate setup before simulation.
}
