% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.4-plot-bioenergetic.R
\name{plot_bio_dashboard}
\alias{plot_bio_dashboard}
\title{Plot Bioenergetic dashboard}
\usage{
plot_bio_dashboard(bio_obj, colors = "blue", title = NULL, ...)
}
\description{
Plot Bioenergetic dashboard
}
\keyword{internal}
