% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-fb4.R
\name{plot_bio_temperature}
\alias{plot_bio_temperature}
\title{Plot Bioenergetic temperature}
\usage{
plot_bio_temperature(bio_obj, colors = "red", ...)
}
\description{
Plot Bioenergetic temperature
}
\keyword{internal}
