% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{plot_body_composition_by_size}
\alias{plot_body_composition_by_size}
\title{Graficar composición corporal por tamaño}
\usage{
plot_body_composition_by_size(analysis_data, plot_type = "fractions")
}
\arguments{
\item{analysis_data}{Output de analyze_composition_by_size}

\item{plot_type}{Tipo de gráfico ("fractions", "absolute", "energy")}
}
\value{
Gráfico de composición corporal
}
\description{
Crea gráficos de cómo cambia la composición con el tamaño
}
