% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.3-plot-analysis.R
\name{plot_bootstrap_distributions}
\alias{plot_bootstrap_distributions}
\title{Plot bootstrap distributions}
\usage{
plot_bootstrap_distributions(fb4_result, colors)
}
\arguments{
\item{fb4_result}{Bootstrap result object}

\item{colors}{Color scheme}
}
\description{
Internal function for bootstrap distribution plotting.
}
\keyword{internal}
