% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{plot_composition_barplot}
\alias{plot_composition_barplot}
\title{Crear gráfico de barras de composición corporal}
\usage{
plot_composition_barplot(compositions, labels = NULL)
}
\arguments{
\item{compositions}{Lista de composiciones o output de compare_body_compositions}

\item{labels}{Etiquetas para cada composición}
}
\value{
Gráfico de barras apiladas
}
\description{
Gráfico de barras apiladas mostrando composición corporal
}
