% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.2-plot-daily.R
\name{plot_consumption}
\alias{plot_consumption}
\title{Plot consumption patterns}
\usage{
plot_consumption(fb4_result, show_cumulative = TRUE, colors = "green", ...)
}
\arguments{
\item{fb4_result}{FB4 result object}

\item{show_cumulative}{Show cumulative consumption subplot}

\item{colors}{Color scheme}

\item{...}{Additional arguments}
}
\value{
NULL (creates plot)
}
\description{
Creates consumption rate plots with optional cumulative consumption.
}
\keyword{internal}
