% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_consumption_components}
\alias{plot_consumption_components}
\title{Plot consumption components (rate and total by diet items)}
\usage{
plot_consumption_components(fb4_result, show_diet_breakdown = TRUE)
}
\arguments{
\item{fb4_result}{FB4 result object}

\item{show_diet_breakdown}{Show consumption by diet items}
}
\description{
Plot consumption components (rate and total by diet items)
}
