% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_consumption_rates}
\alias{plot_consumption_rates}
\title{Graficar tasas de consumo}
\usage{
plot_consumption_rates(daily_data, main = "Tasas de Consumo", ...)
}
\arguments{
\item{daily_data}{Datos diarios}

\item{main}{Título del gráfico}

\item{...}{Argumentos adicionales}
}
\description{
Graficar tasas de consumo
}
\keyword{internal}
