% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_diet_consumption}
\alias{plot_diet_consumption}
\title{Plot diet consumption breakdown}
\usage{
plot_diet_consumption(fb4_result)
}
\arguments{
\item{fb4_result}{FB4 result object}
}
\description{
Plot diet consumption breakdown
}
\keyword{internal}
