% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_energy_budget}
\alias{plot_energy_budget}
\title{Graficar presupuesto energético}
\usage{
plot_energy_budget(daily_data, main = "Presupuesto Energético", ...)
}
\arguments{
\item{daily_data}{Datos diarios}

\item{main}{Título del gráfico}

\item{...}{Argumentos adicionales}
}
\description{
Graficar presupuesto energético
}
\keyword{internal}
