% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_energy_components}
\alias{plot_energy_components}
\title{Plot energy components}
\usage{
plot_energy_components(
  fb4_result,
  components = c("Consumption_energy", "Respiration", "Net_energy")
)
}
\arguments{
\item{fb4_result}{FB4 result object}

\item{components}{Energy components to plot}
}
\description{
Plot energy components
}
