% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.2-plot-daily.R
\name{plot_growth}
\alias{plot_growth}
\title{Plot growth trajectory}
\usage{
plot_growth(fb4_result, show_rate = TRUE, colors = "blue", smooth = FALSE, ...)
}
\arguments{
\item{fb4_result}{FB4 result object}

\item{show_rate}{Show growth rate subplot}

\item{colors}{Color scheme}

\item{smooth}{Add smoothed trend}

\item{...}{Additional arguments}
}
\value{
NULL (creates plot)
}
\description{
Creates growth trajectory plots with optional growth rate subplot.
}
\keyword{internal}
