% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_growth_components}
\alias{plot_growth_components}
\title{Plot growth components (weight gain and cumulative)}
\usage{
plot_growth_components(fb4_result, show_cumulative = TRUE)
}
\arguments{
\item{fb4_result}{FB4 result object}

\item{show_cumulative}{Show cumulative growth}
}
\description{
Plot growth components (weight gain and cumulative)
}
