% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-fb4.R
\name{plot_growth_temperature_sensitivity}
\alias{plot_growth_temperature_sensitivity}
\title{Plot sensitivity analysis}
\usage{
plot_growth_temperature_sensitivity(
  sensitivity_data,
  temperatures = seq(5, 20, by = 2),
  feeding_levels = c(0.5, 0.75, 1),
  species = NULL,
  ylim = NULL,
  xlim = NULL,
  colors = "grayscale",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{temperatures}{Temperature values to test}

\item{feeding_levels}{Feeding levels to test}

\item{colors}{Color scheme}

\item{verbose}{Show progress messages}

\item{...}{Additional arguments}

\item{bio_obj}{Bioenergetic object}
}
\value{
NULL (creates plot)
}
\description{
Creates sensitivity analysis plots for temperature and feeding effects.
}
