% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.3-plot-analysis.R
\name{plot_hierarchical_distributions}
\alias{plot_hierarchical_distributions}
\title{Plot hierarchical distributions}
\usage{
plot_hierarchical_distributions(fb4_result, colors, show_individuals)
}
\arguments{
\item{fb4_result}{Hierarchical result object}

\item{colors}{Color scheme}

\item{show_individuals}{Show individual estimates}
}
\description{
Internal function for hierarchical distribution plotting.
}
\keyword{internal}
