% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-fb4.R, R/16.3-plot-analysis.R
\name{plot_hierarchical_uncertainty}
\alias{plot_hierarchical_uncertainty}
\title{Plot hierarchical uncertainty}
\usage{
plot_hierarchical_uncertainty(fb4_result, parameters, colors, add_ci_text)

plot_hierarchical_uncertainty(fb4_result, parameters, colors, add_ci_text)
}
\arguments{
\item{fb4_result}{Hierarchical result object}

\item{parameters}{Parameters to plot}

\item{colors}{Color scheme}

\item{add_ci_text}{Add confidence interval text}
}
\description{
Internal function for hierarchical uncertainty plotting.
}
\keyword{internal}
