% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-fb4.R, R/16.3-plot-analysis.R
\name{plot_mle_uncertainty}
\alias{plot_mle_uncertainty}
\title{Plot MLE uncertainty}
\usage{
plot_mle_uncertainty(fb4_result, parameters, colors, add_ci_text)

plot_mle_uncertainty(fb4_result, parameters, colors, add_ci_text)
}
\arguments{
\item{fb4_result}{MLE result object}

\item{parameters}{Parameters to plot}

\item{colors}{Color scheme}

\item{add_ci_text}{Add confidence interval text}
}
\description{
Internal function for MLE uncertainty plotting.
}
\keyword{internal}
