% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{plot_nutrient_balance_barplot}
\alias{plot_nutrient_balance_barplot}
\title{Crear gráfico de barras de balance de nutrientes}
\usage{
plot_nutrient_balance_barplot(nitrogen_fluxes, phosphorus_fluxes)
}
\arguments{
\item{nitrogen_fluxes}{Vector resultado de nitrogen_allocation}

\item{phosphorus_fluxes}{Vector resultado de phosphorous_allocation}
}
\value{
Gráfico de barras del balance de nutrientes
}
\description{
Gráfico de barras mostrando el destino de N y P consumidos
}
