% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{plot_nutrient_dynamics}
\alias{plot_nutrient_dynamics}
\title{Graficar dinámica de nutrientes}
\usage{
plot_nutrient_dynamics(simulation_results, plot_type = "fluxes")
}
\arguments{
\item{simulation_results}{Output de simulate_nutrient_dynamics}

\item{plot_type}{Tipo de gráfico ("fluxes", "ratios", "cumulative")}
}
\value{
Gráfico de dinámica de nutrientes
}
\description{
Crea gráficos de los flujos de nutrientes en el tiempo
}
