% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_population_dynamics}
\alias{plot_population_dynamics}
\title{Graficar dinámicas poblacionales}
\usage{
plot_population_dynamics(daily_data, main = "Dinámicas Poblacionales", ...)
}
\arguments{
\item{daily_data}{Datos diarios}

\item{main}{Título del gráfico}

\item{...}{Argumentos adicionales}
}
\description{
Graficar dinámicas poblacionales
}
\keyword{internal}
