% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{plot_predator_energy_density}
\alias{plot_predator_energy_density}
\title{Graficar densidad energética vs peso}
\usage{
plot_predator_energy_density(
  PREDEDEQ,
  weight_range = c(1, 500),
  n_points = 100,
  ...
)
}
\arguments{
\item{PREDEDEQ}{Ecuación a graficar}

\item{weight_range}{Rango de pesos (vector de 2 elementos)}

\item{n_points}{Número de puntos para graficar}

\item{...}{Parámetros específicos de la ecuación}
}
\value{
Gráfico de densidad energética
}
\description{
Crea gráfico de la relación entre peso y densidad energética
}
