% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-fb4.R
\name{plot_temperature}
\alias{plot_temperature}
\title{Plot temperature profile}
\usage{
plot_temperature(
  fb4_result,
  show_correlation = TRUE,
  colors = "red",
  smooth = TRUE,
  ...
)
}
\arguments{
\item{fb4_result}{FB4 result object}

\item{show_correlation}{Show temperature-consumption relationship}

\item{colors}{Color scheme}

\item{smooth}{Add smoothed trend}

\item{...}{Additional arguments}
}
\value{
NULL (creates plot)
}
\description{
Creates temperature plots with optional consumption correlation.
}
\keyword{internal}
