% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_temperature_profile}
\alias{plot_temperature_profile}
\title{Plot temperature profile}
\usage{
plot_temperature_profile(fb4_result, add_smooth = TRUE)
}
\arguments{
\item{fb4_result}{FB4 result object}

\item{add_smooth}{Add smoothed trend line}
}
\description{
Plot temperature profile
}
