% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{predator_energy_eq2}
\alias{predator_energy_eq2}
\title{Densidad energética lineal por segmentos (Ecuación 2)}
\usage{
predator_energy_eq2(weight, Alpha1, Beta1, Alpha2, Beta2, Cutoff)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{Alpha1}{Intercepto para primer segmento}

\item{Beta1}{Pendiente para primer segmento}

\item{Alpha2}{Intercepto para segundo segmento}

\item{Beta2}{Pendiente para segundo segmento}

\item{Cutoff}{Peso de corte entre segmentos}
}
\value{
Densidad energética (J/g)
}
\description{
Densidad energética lineal por segmentos (Ecuación 2)
}
\keyword{internal}
