% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{predator_energy_eq3}
\alias{predator_energy_eq3}
\title{Densidad energética función potencia (Ecuación 3)}
\usage{
predator_energy_eq3(weight, Alpha1, Beta1)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{Alpha1}{Coeficiente multiplicativo}

\item{Beta1}{Exponente}
}
\value{
Densidad energética (J/g)
}
\description{
Densidad energética función potencia (Ecuación 3)
}
\keyword{internal}
