% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{predator_energy_power_function}
\alias{predator_energy_power_function}
\title{Densidad energética función potencia}
\usage{
predator_energy_power_function(W, alpha1, beta1)
}
\arguments{
\item{W}{Peso del pez (g)}

\item{alpha1}{Coeficiente multiplicativo}

\item{beta1}{Exponente}
}
\value{
Densidad energética (J/g)
}
\description{
Calcula densidad energética usando ED = alpha1 * W^beta1
}
\examples{
# Densidad energética que aumenta con el peso
predator_energy_power_function(W = 50, alpha1 = 4000, beta1 = 0.1)

# Densidad energética constante (beta1 = 0)
predator_energy_power_function(W = 50, alpha1 = 4500, beta1 = 0)
}
