% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{predict.Bioenergetic}
\alias{predict.Bioenergetic}
\title{Método predict para objetos Bioenergetic}
\usage{
\method{predict}{Bioenergetic}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Objeto Bioenergetic ajustado}

\item{newdata}{Nuevos datos ambientales (opcional)}

\item{...}{Argumentos adicionales}
}
\value{
Predicciones del modelo
}
\description{
Método predict para objetos Bioenergetic
}
