% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_composition_params}
\alias{process_composition_params}
\title{Process body composition parameters}
\usage{
process_composition_params(composition_params)
}
\arguments{
\item{composition_params}{Raw composition parameters}
}
\value{
Processed composition parameters
}
\description{
Process body composition parameters
}
