% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{process_contaminant_data_simple}
\alias{process_contaminant_data_simple}
\title{Procesar datos de contaminantes simplificado}
\usage{
process_contaminant_data_simple(contaminant_data, target_days, prey_names)
}
\arguments{
\item{contaminant_data}{Lista con datos de contaminantes}

\item{target_days}{Vector de días objetivo}

\item{prey_names}{Vector con nombres de presas}
}
\value{
Lista con datos de contaminantes procesados
}
\description{
Procesar datos de contaminantes simplificado
}
\keyword{internal}
