% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_egestion_params}
\alias{process_egestion_params}
\title{Process egestion parameters}
\usage{
process_egestion_params(egestion_params)
}
\arguments{
\item{egestion_params}{Raw egestion parameters}
}
\value{
Processed egestion parameters
}
\description{
Process egestion parameters
}
