% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R, R/12-fb4-main.R
\name{process_input_data}
\alias{process_input_data}
\title{Procesar datos de entrada para el modelo}
\usage{
process_input_data(
  temperature_data,
  diet_data,
  prey_energy_data,
  first_day,
  last_day,
  mortality_data = NULL,
  reproduction_data = NULL,
  contaminant_data = NULL,
  nutrient_data = NULL
)

process_input_data(
  temperature_data,
  diet_data,
  prey_energy_data,
  first_day,
  last_day,
  mortality_data = NULL,
  reproduction_data = NULL,
  contaminant_data = NULL,
  nutrient_data = NULL
)
}
\arguments{
\item{temperature_data}{Data frame con columnas Day y Temperature}

\item{diet_data}{Data frame con proporciones de dieta}

\item{prey_energy_data}{Data frame con energías de presas}

\item{first_day}{Primer día de simulación}

\item{last_day}{Último día de simulación}
}
\value{
Lista con datos procesados
}
\description{
Procesar datos de entrada para el modelo

Procesar datos de entrada para FB4
}
\keyword{internal}
