% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_mortality_params}
\alias{process_mortality_params}
\title{Process mortality parameters}
\usage{
process_mortality_params(mortality_params)
}
\arguments{
\item{mortality_params}{Raw mortality parameters}
}
\value{
Processed mortality parameters
}
\description{
Process mortality parameters
}
