% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{process_nutrient_data_simple}
\alias{process_nutrient_data_simple}
\title{Procesar datos de nutrientes simplificado}
\usage{
process_nutrient_data_simple(nutrient_data, target_days, prey_names)
}
\arguments{
\item{nutrient_data}{Lista con datos de nutrientes}

\item{target_days}{Vector de días objetivo}

\item{prey_names}{Vector con nombres de presas}
}
\value{
Lista con datos de nutrientes procesados
}
\description{
Procesar datos de nutrientes simplificado
}
\keyword{internal}
