% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_predator_energy_data}
\alias{process_predator_energy_data}
\title{Process predator energy data for equation 1}
\usage{
process_predator_energy_data(predator_params)
}
\arguments{
\item{predator_params}{Predator parameters}
}
\value{
Processed parameters with ED_data vector
}
\description{
Process predator energy data for equation 1
}
\keyword{internal}
