% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-data-processing.R
\name{process_reproduction_data}
\alias{process_reproduction_data}
\title{Process reproduction data}
\usage{
process_reproduction_data(bio_obj, target_days)
}
\arguments{
\item{bio_obj}{Bioenergetic object}

\item{target_days}{Target simulation days}
}
\value{
Vector of reproduction fractions for each day
}
\description{
Process reproduction data
}
\keyword{internal}
