% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_reproduction_pattern}
\alias{process_reproduction_pattern}
\title{Process reproduction pattern}
\usage{
process_reproduction_pattern(mortality_params)
}
\description{
Process reproduction pattern
}
\keyword{internal}
