% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_respiration_params}
\alias{process_respiration_params}
\title{Process respiration parameters}
\usage{
process_respiration_params(
  respiration_params,
  activity_params = NULL,
  sda_params = NULL
)
}
\arguments{
\item{respiration_params}{Raw respiration parameters}

\item{activity_params}{Activity parameters (required for REQ=1)}

\item{sda_params}{SDA parameters}
}
\value{
Processed respiration parameters with derived values
}
\description{
Process respiration parameters
}
