% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.3-analysis-sensitivity.R
\name{process_temperature}
\alias{process_temperature}
\title{Process single temperature with all p-values}
\usage{
process_temperature(
  temp,
  p_values,
  processed_data,
  simulation_days,
  oxycal,
  verbose
)
}
\arguments{
\item{temp}{Temperature value}

\item{p_values}{Vector of p-values to test}

\item{processed_data}{Prepared simulation data}

\item{simulation_days}{Number of simulation days}

\item{oxycal}{Oxycalorific coefficient}

\item{verbose}{Show progress}
}
\value{
Data frame with results for this temperature
}
\description{
Process single temperature with all p-values
}
\keyword{internal}
