% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{project_population_longterm}
\alias{project_population_longterm}
\title{Proyectar población a largo plazo}
\usage{
project_population_longterm(
  initial_population,
  mortality_rates,
  recruitment_rate = 0,
  projection_days = 365
)
}
\arguments{
\item{initial_population}{Población inicial}

\item{mortality_rates}{Vector de tasas de mortalidad promedio por tipo}

\item{recruitment_rate}{Tasa de reclutamiento (opcional)}

\item{projection_days}{Días a proyectar}
}
\value{
Data frame con proyección poblacional
}
\description{
Proyecta dinámica poblacional usando tasas promedio
}
