% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{prt_msg}
\alias{prt_msg}
\title{Imprimir mensaje de error con información de contexto}
\usage{
prt_msg(run, day, wt, waterT, p, outpt, msg_num, msg_loc)
}
\arguments{
\item{run}{Número de corrida}

\item{day}{Día}

\item{wt}{Peso}

\item{waterT}{Temperatura del agua}

\item{p}{Valor p}

\item{outpt}{Tipo de salida}

\item{msg_num}{Número de mensaje}

\item{msg_loc}{Ubicación del mensaje}
}
\value{
NULL (imprime mensaje)
}
\description{
Imprimir mensaje de error con información de contexto
}
