% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{RACTf1T}
\alias{RACTf1T}
\title{Función de actividad para respiración - Ecuación 1}
\usage{
RACTf1T(W, Temperature, RTL, ACT, RK4, BACT, RK1, RK5, RTO)
}
\arguments{
\item{W}{Peso del pez (g)}

\item{Temperature}{Temperatura del agua (°C)}

\item{RTL}{Temperatura de corte para cambio en relación de actividad}

\item{ACT}{Multiplicador de actividad}

\item{RK4}{Coeficiente de dependencia de masa para velocidad de nado}

\item{BACT}{Coeficiente de dependencia de temperatura para velocidad de nado}

\item{RK1}{Intercepto para velocidad de nado sobre temperatura de corte}

\item{RK5}{Coeficiente de temperatura para velocidad de nado}

\item{RTO}{Temperatura óptima para respiración}
}
\value{
Factor de actividad
}
\description{
Implementa la función de actividad con componente de temperatura
}
