% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-simulation-engine.R
\name{report_simulation_progress}
\alias{report_simulation_progress}
\title{Report simulation progress}
\usage{
report_simulation_progress(
  day,
  n_days,
  current_weight,
  initial_weight,
  effective_p,
  temperature
)
}
\description{
Report simulation progress
}
\keyword{internal}
