% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-simulation-engine.R
\name{report_simulation_summary}
\alias{report_simulation_summary}
\title{Report final simulation summary}
\usage{
report_simulation_summary(
  final_weight,
  initial_weight,
  total_consumption,
  n_days
)
}
\description{
Report final simulation summary
}
\keyword{internal}
