% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{respiration}
\alias{respiration}
\title{Función principal de respiración}
\usage{
respiration(Temperature, W, REQ, RA, RB, ...)
}
\arguments{
\item{Temperature}{Temperatura del agua (°C)}

\item{W}{Peso del pez (g)}

\item{REQ}{Número de ecuación de respiración (1-2)}

\item{RA}{Peso específico de oxígeno consumido por pez de 1g a 0°C}

\item{RB}{Pendiente de función masa alométrica para metabolismo estándar}

\item{...}{Parámetros adicionales según la ecuación seleccionada}
}
\value{
Respiración (g O2/g/día)
}
\description{
Calcula la respiración basada en temperatura, peso y ecuación seleccionada
}
\examples{
# Usando ecuación 1
respiration(Temperature = 20, W = 100, REQ = 1,
            RA = 0.0038, RB = -0.355, RQ = 0.08,
            RTL = 24, ACT = 1, RK4 = 0.13, BACT = 0.08,
            RK1 = 1, RK5 = 0.13, RTO = 0.030)
}
