% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{respiration_temp_eq1}
\alias{respiration_temp_eq1}
\title{Función de temperatura para respiración - Ecuación 1}
\usage{
respiration_temp_eq1(temperature, RQ)
}
\arguments{
\item{temperature}{Temperatura del agua (°C)}

\item{RQ}{Tasa Q10 para temperaturas bajas}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuación de temperatura 1 (exponencial simple)
}
\keyword{internal}
