% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{Rf1T}
\alias{Rf1T}
\title{Función de temperatura para respiración - Ecuación 1}
\usage{
Rf1T(Temperature, RQ)
}
\arguments{
\item{Temperature}{Temperatura del agua (°C)}

\item{RQ}{Tasa Q10 para temperaturas bajas}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuación de temperatura 1 para respiración (Hanson et al. 1997; Stewart et al. 1983)
}
