% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{Rf2T}
\alias{Rf2T}
\title{Función de temperatura para respiración - Ecuación 2}
\usage{
Rf2T(Temperature, RTM, RTO, RX)
}
\arguments{
\item{Temperature}{Temperatura del agua (°C)}

\item{RTM}{Temperatura máxima (letal)}

\item{RTO}{Temperatura óptima para respiración}

\item{RX}{Parámetro calculado}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuación de temperatura 2 para respiración (Hanson et al. 1997; Kitchell et al. 1977)
}
