% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.0-run-fb4-orchestrator.R
\name{run_fb4.default}
\alias{run_fb4.default}
\title{Default method for run_fb4 (error)}
\usage{
\method{run_fb4}{default}(x, ...)
}
\arguments{
\item{x}{Object of unsupported class}

\item{...}{Additional arguments}
}
\value{
No return value. Called for its side effect of stopping execution
  with an informative error message when \code{x} is not a
  \code{Bioenergetic} object.
}
\description{
Default method that provides helpful error message.
}
