% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4_direct}
\alias{run_fb4_direct}
\title{Ejecutar FB4 directo}
\usage{
run_fb4_direct(
  species_params,
  initial_weight,
  fit_to,
  fit_value,
  processed_data,
  model_options,
  oxycal,
  population_size,
  output_daily,
  progress_callback = NULL
)
}
\description{
Ejecutar FB4 directo
}
\keyword{internal}
