% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4_simulation_complete}
\alias{run_fb4_simulation_complete}
\title{Complete FB4 Simulation}
\usage{
run_fb4_simulation_complete(
  initial_weight,
  p_value,
  species_params,
  processed_data,
  model_options,
  oxycal = 13560,
  output_daily = TRUE
)
}
\description{
Complete FB4 Simulation
}
\keyword{internal}
