% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4_simulation_with_ration}
\alias{run_fb4_simulation_with_ration}
\title{FB4 Simulation with Specific Ration}
\usage{
run_fb4_simulation_with_ration(
  initial_weight,
  ration_type,
  ration_value,
  species_params,
  processed_data,
  model_options,
  oxycal = 13560,
  output_daily = TRUE
)
}
\description{
FB4 Simulation with Specific Ration
}
\keyword{internal}
