% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{run_fb4_with_fitting}
\alias{run_fb4_with_fitting}
\title{Ejecutar modelo FB4 con ajuste de p-value}
\usage{
run_fb4_with_fitting(
  species_params,
  initial_weight,
  fit_to,
  fit_value,
  processed_data,
  model_options,
  oxycal,
  population_size,
  output_daily,
  progress_callback
)
}
\arguments{
\item{species_params}{Parámetros de la especie}

\item{initial_weight}{Peso inicial}

\item{fit_to}{Tipo de ajuste}

\item{fit_value}{Valor objetivo}

\item{processed_data}{Datos procesados}

\item{model_options}{Opciones del modelo}

\item{oxycal}{Coeficiente oxicalórico}

\item{population_size}{Tamaño de población}

\item{output_daily}{Incluir salida diaria}

\item{progress_callback}{Función callback para progreso}
}
\value{
Resultado de la simulación
}
\description{
Ejecutar modelo FB4 con ajuste de p-value
}
\keyword{internal}
