% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4_with_p_value}
\alias{run_fb4_with_p_value}
\title{Execute FB4 Simulation with Specific P-Value}
\usage{
run_fb4_with_p_value(
  species_params,
  initial_weight,
  p_value,
  processed_data,
  model_options,
  oxycal = 13560,
  output_daily = TRUE
)
}
\description{
Execute FB4 Simulation with Specific P-Value
}
\keyword{internal}
