% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4_with_ration_percent}
\alias{run_fb4_with_ration_percent}
\title{Execute FB4 with Ration as Percentage of Body Weight}
\usage{
run_fb4_with_ration_percent(
  species_params,
  initial_weight,
  ration_percent,
  processed_data,
  model_options,
  oxycal = 13560,
  output_daily = TRUE
)
}
\description{
Execute FB4 with Ration as Percentage of Body Weight
}
\keyword{internal}
