% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{run_with_p_value}
\alias{run_with_p_value}
\title{Ejecutar modelo con p-value específico}
\usage{
run_with_p_value(bioenergetic_obj, p_value, return_daily = TRUE)
}
\arguments{
\item{bioenergetic_obj}{Objeto de clase Bioenergetic}

\item{p_value}{Valor de p a usar}

\item{return_daily}{Retornar datos diarios}
}
\value{
Resultado de la simulación
}
\description{
Ejecuta el modelo bioenergético con un p-value determinado
}
