% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{save_daily_results}
\alias{save_daily_results}
\title{Guardar resultados diarios}
\usage{
save_daily_results(
  daily_output,
  day,
  day_data,
  current_weight,
  final_weight,
  current_population,
  consumption_gg,
  consumption_energy,
  egestion_energy,
  excretion_energy,
  respiration_energy,
  sda_energy,
  growth_energy,
  weight_gain,
  spawn_energy,
  predator_energy_start,
  predator_energy_end,
  mean_prey_energy,
  consumption_by_prey,
  contaminant_results,
  nutrient_results,
  processed_data
)
}
\arguments{
\item{daily_output}{Data frame de salida diaria}

\item{day}{Día actual}

\item{day_data}{Datos del día}

\item{current_weight}{Peso actual}

\item{final_weight}{Peso final}

\item{current_population}{Población actual}

\item{consumption_gg}{Consumo específico}

\item{consumption_energy}{Consumo energético}

\item{egestion_energy}{Egestion}

\item{excretion_energy}{Excreción}

\item{respiration_energy}{Respiración}

\item{sda_energy}{SDA}

\item{growth_energy}{Energía de crecimiento}

\item{weight_gain}{Ganancia de peso}

\item{spawn_energy}{Energía de reproducción}

\item{predator_energy_start}{Densidad energética inicial}

\item{predator_energy_end}{Densidad energética final}

\item{mean_prey_energy}{Energía media de presas}

\item{consumption_by_prey}{Consumo por tipo de presa}

\item{contaminant_results}{Resultados de contaminantes}

\item{nutrient_results}{Resultados de nutrientes}

\item{processed_data}{Datos procesados}
}
\value{
Data frame actualizado
}
\description{
Guardar resultados diarios
}
\keyword{internal}
