% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_diet.Bioenergetic}
\alias{set_diet.Bioenergetic}
\title{Set Diet Data for Bioenergetic Objects}
\usage{
\method{set_diet}{Bioenergetic}(x, diet_proportions, prey_energies, indigestible_prey = NULL)
}
\arguments{
\item{x}{Bioenergetic object}

\item{diet_proportions}{Data frame with daily diet proportions}

\item{prey_energies}{Data frame with daily prey energy densities}
}
\value{
Updated Bioenergetic object with new diet data
}
\description{
Updates the diet data component of a Bioenergetic object with
new diet composition and prey energy information.
}
\details{
Both data frames must:
\itemize{
  \item{Contain a "Day" column with matching values}
  \item{Have identical prey species columns}
  \item{Use consistent prey naming between datasets}
}

Diet proportions should sum to approximately 1.0 for each day.
Prey energies should be positive values in J/g.
}
\examples{
\dontrun{
bio_obj <- new_bioenergetic("Salmo salar", 10)
diet_props <- data.frame(Day = 1:365, fish = 0.6, zooplankton = 0.4)
prey_energy <- data.frame(Day = 1:365, fish = 5000, zooplankton = 3000)
bio_obj <- set_diet(bio_obj, diet_props, prey_energy)
}
}
