% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_environment.Bioenergetic}
\alias{set_environment.Bioenergetic}
\title{Set Environmental Data for Bioenergetic Objects}
\usage{
\method{set_environment}{Bioenergetic}(x, temperature_data)
}
\arguments{
\item{x}{Bioenergetic object}

\item{temperature_data}{Data frame with Day and Temperature columns}
}
\value{
Updated Bioenergetic object with new environmental data
}
\description{
Updates the environmental data component of a Bioenergetic object
with new temperature information.
}
\details{
The temperature data frame must contain:
\itemize{
  \item{Day: }{Numeric sequence of simulation days}
  \item{Temperature: }{Temperature values in °C}
}

Setting new environmental data resets the fitted status and clears
any existing results, as the model needs to be re-run.
}
\examples{
\dontrun{
bio_obj <- new_bioenergetic("Salmo salar", 10)
temp_data <- data.frame(Day = 1:100, Temperature = 15 + 5*sin(1:100/50))
bio_obj <- set_environment(bio_obj, temp_data)
}
}
