% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_options}
\alias{set_options}
\title{Configurar opciones del modelo}
\usage{
set_options(x, ...)
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{...}{Opciones del modelo como argumentos nombrados}
}
\value{
Objeto Bioenergetic modificado
}
\description{
Configurar opciones del modelo
}
