% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0-bioenergetic-classes.R
\name{set_parameter_value}
\alias{set_parameter_value}
\title{Set Parameter Value in Species Parameters}
\usage{
set_parameter_value(params, param, value)
}
\arguments{
\item{params}{Species parameters list}

\item{param}{Parameter name to set}

\item{value}{New parameter value}
}
\value{
The \code{params} list with \code{params[[category]][[param]]}
  replaced by \code{value}, where \code{category} is the first category
  in which \code{param} is found. Throws an error if \code{param} is not
  found in any category.
}
\description{
Sets a specific parameter value in species parameter lists,
automatically finding the correct category.
}
\examples{
sp <- list(consumption = list(CA = 0.303, CB = -0.275))
updated <- set_parameter_value(sp, "CA", 0.350)
updated$consumption$CA
}
