% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_simulation}
\alias{set_simulation}
\title{Configurar parámetros de simulación}
\usage{
set_simulation(x, ...)
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{...}{Parámetros de simulación}
}
\value{
Objeto Bioenergetic modificado
}
\description{
Configurar parámetros de simulación
}
