% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0-bioenergetic-classes.R
\name{set_simulation_settings}
\alias{set_simulation_settings}
\title{Set Simulation Settings for Bioenergetic Objects}
\usage{
set_simulation_settings(x, initial_weight = NULL, duration = NULL)
}
\arguments{
\item{x}{Bioenergetic object}

\item{initial_weight}{Initial weight in grams}

\item{duration}{Simulation duration in days (auto-detected if NULL)}
}
\value{
Updated Bioenergetic object with new simulation settings
}
\description{
Updates the simulation configuration of a Bioenergetic object.
}
