% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{setup_model_options}
\alias{setup_model_options}
\title{Configurar opciones del modelo}
\usage{
setup_model_options(
  calc_mortality = FALSE,
  calc_reproduction = FALSE,
  calc_contaminant = FALSE,
  calc_nutrient = FALSE,
  ...
)
}
\description{
Configurar opciones del modelo
}
\keyword{internal}
