% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{setup_model_options_from_bioenergetic}
\alias{setup_model_options_from_bioenergetic}
\title{Configurar opciones del modelo desde objeto Bioenergetic}
\usage{
setup_model_options_from_bioenergetic(bio_obj, ...)
}
\description{
Configurar opciones del modelo desde objeto Bioenergetic
}
\keyword{internal}
