% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.1-plot-core.R
\name{setup_save_device}
\alias{setup_save_device}
\title{Setup graphics device for saving}
\usage{
setup_save_device(save_path)
}
\arguments{
\item{save_path}{Path to save file (.png or .pdf)}
}
\description{
Opens appropriate graphics device based on file extension.
}
\keyword{internal}
