% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.3-result-builders.R
\name{should_mark_as_fitted}
\alias{should_mark_as_fitted}
\title{Determine if bio_obj should be marked as fitted}
\usage{
should_mark_as_fitted(result)
}
\description{
Determine if bio_obj should be marked as fitted
}
\keyword{internal}
