% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{simulate}
\alias{simulate}
\title{Simular modelo bioenergético}
\usage{
simulate(x, ...)
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{...}{Argumentos adicionales pasados al simulador}
}
\value{
Objeto Bioenergetic con resultados
}
\description{
Simular modelo bioenergético
}
