% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{solve_linear_segments_weight}
\alias{solve_linear_segments_weight}
\title{Resolver peso final para función lineal por segmentos}
\usage{
solve_linear_segments_weight(
  initial_weight,
  net_energy,
  initial_body_energy,
  alpha1,
  beta1,
  alpha2,
  beta2,
  cutoff
)
}
\arguments{
\item{initial_weight}{Peso inicial}

\item{net_energy}{Energía neta disponible}

\item{initial_body_energy}{Energía corporal inicial}

\item{alpha1}{Parámetros de la función lineal}

\item{beta1}{Parámetros de la función lineal}

\item{alpha2}{Parámetros de la función lineal}

\item{beta2}{Parámetros de la función lineal}

\item{cutoff}{Peso de corte}
}
\value{
Peso final
}
\description{
Función auxiliar para calcular peso final con PREDEDEQ = 2
}
\keyword{internal}
