% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{solve_weight_linear_segments}
\alias{solve_weight_linear_segments}
\title{Solve weight for linear segments (Low-level)}
\usage{
solve_weight_linear_segments(
  available_energy,
  Alpha1,
  Beta1,
  Alpha2,
  Beta2,
  Cutoff
)
}
\arguments{
\item{available_energy}{Available energy (J)}

\item{Alpha1}{Intercept for first size segment}

\item{Beta1}{Slope for first size segment}

\item{Alpha2}{Intercept for second size segment}

\item{Beta2}{Slope for second size segment}

\item{Cutoff}{Weight cutoff between segments (g)}
}
\value{
Final weight (g)
}
\description{
Solves final weight for piecewise linear function (PREDEDEQ = 2)
}
\keyword{internal}
