% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{solve_weight_power_function}
\alias{solve_weight_power_function}
\title{Solve weight for power function (Low-level)}
\usage{
solve_weight_power_function(initial_weight, net_energy, Alpha1, Beta1)
}
\arguments{
\item{initial_weight}{Initial weight}

\item{net_energy}{Net energy}

\item{Alpha1}{Coefficient}

\item{Beta1}{Exponent}
}
\value{
Final weight
}
\description{
Solves final weight for power function relationship (PREDEDEQ = 3)
}
\keyword{internal}
