% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{SpDynAct}
\alias{SpDynAct}
\title{Función de acción dinámica específica}
\usage{
SpDynAct(C, Eg, SDA)
}
\arguments{
\item{C}{Consumo (J/g)}

\item{Eg}{Egestion (J/g)}

\item{SDA}{Coeficiente de acción dinámica específica}
}
\value{
SDA (J/g)
}
\description{
Calcula la acción dinámica específica (SDA)
}
