% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{species_info}
\alias{species_info}
\title{Obtener información de una especie}
\usage{
species_info(species, fish4_db = NULL)
}
\arguments{
\item{species}{Nombre científico de la especie}

\item{fish4_db}{Base de datos fish4_parameters (opcional)}
}
\value{
Lista con información de la especie
}
\description{
Obtener información de una especie
}
