% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{standardize_vector_length}
\alias{standardize_vector_length}
\title{Standardize vector length (Utility)}
\usage{
standardize_vector_length(vector, target_length, param_name)
}
\arguments{
\item{vector}{Input vector}

\item{target_length}{Target length}

\item{param_name}{Parameter name for error messages}
}
\value{
Vector with correct length
}
\description{
Ensures vector has correct length for calculations
}
\keyword{internal}
