% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{summarize_fb4_fit_result}
\alias{summarize_fb4_fit_result}
\title{Generar resumen detallado de ajuste FB4}
\usage{
summarize_fb4_fit_result(fit_result)
}
\arguments{
\item{fit_result}{Resultado de función de ajuste FB4}
}
\value{
Data frame con resumen detallado
}
\description{
Generar resumen detallado de ajuste FB4
}
