% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{summarize_fit_result}
\alias{summarize_fit_result}
\title{Generar resumen de ajuste}
\usage{
summarize_fit_result(fit_result)
}
\arguments{
\item{fit_result}{Resultado de función de ajuste}
}
\value{
Data frame con resumen
}
\description{
Generar resumen de ajuste
}
