% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{summarize_simulation_result}
\alias{summarize_simulation_result}
\title{Generar resumen de simulación}
\usage{
summarize_simulation_result(simulation_result)
}
\arguments{
\item{simulation_result}{Resultado de run_bioenergetic_simulation}
}
\value{
Data frame con resumen
}
\description{
Generar resumen de simulación
}
