% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.1-bioenergetic-methods.R
\name{summary.Bioenergetic}
\alias{summary.Bioenergetic}
\title{Summary Method for Bioenergetic Objects (Improved)}
\usage{
\method{summary}{Bioenergetic}(object, ...)
}
\arguments{
\item{object}{Bioenergetic object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Summary Method for Bioenergetic Objects (Improved)
}
