% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.3-analysis-sensitivity.R
\name{test_metric_differences}
\alias{test_metric_differences}
\title{Test differences between metric estimates}
\usage{
test_metric_differences(estimates, se_values, scenario_names)
}
\arguments{
\item{estimates}{Vector of estimates}

\item{se_values}{Vector of standard errors}

\item{scenario_names}{Vector of scenario names}
}
\value{
List with test results
}
\description{
Tests for significant differences between metric estimates using
confidence interval overlap method.
}
\keyword{internal}
