% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-data-processing.R
\name{transform_to_tmb_basic}
\alias{transform_to_tmb_basic}
\title{Transform simulation data to TMB basic format}
\usage{
transform_to_tmb_basic(simulation_data, observed_weights)
}
\description{
Transform simulation data to TMB basic format
}
\keyword{internal}
