% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-data-processing.R
\name{transform_to_tmb_hierarchical}
\alias{transform_to_tmb_hierarchical}
\title{Transform simulation data to TMB hierarchical format}
\usage{
transform_to_tmb_hierarchical(
  simulation_data,
  individual_data = NULL,
  covariates = NULL
)
}
\arguments{
\item{simulation_data}{Processed simulation data}

\item{individual_data}{Data frame with columns: individual_id, initial_weight and observed_weight}

\item{covariates}{Optional covariate matrix or data frame or choose a column of individual_data}
}
\description{
Transform simulation data to TMB hierarchical format
}
\keyword{internal}
