% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.4-main-validators.R
\name{validate_backend_compatibility}
\alias{validate_backend_compatibility}
\title{Validate backend compatibility}
\usage{
validate_backend_compatibility(strategy, backend, verbose = FALSE)
}
\description{
Validate backend compatibility
}
\keyword{internal}
