% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{validate_bioenergetic}
\alias{validate_bioenergetic}
\title{Validar objeto Bioenergetic}
\usage{
validate_bioenergetic(x)
}
\arguments{
\item{x}{Objeto Bioenergetic}
}
\value{
TRUE si es válido, error si no
}
\description{
Validar objeto Bioenergetic
}
\keyword{internal}
