% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{validate_bioenergetic_object}
\alias{validate_bioenergetic_object}
\title{Validar objeto Bioenergetic para FB4}
\usage{
validate_bioenergetic_object(bio_obj)
}
\description{
Validar objeto Bioenergetic para FB4
}
\keyword{internal}
