% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_consumption_params}
\alias{validate_consumption_params}
\title{Validar par치metros de consumo}
\usage{
validate_consumption_params(consumption_params)
}
\arguments{
\item{consumption_params}{Lista con par치metros de consumo}
}
\value{
Lista con resultados de validaci칩n
}
\description{
Validar par치metros de consumo
}
\keyword{internal}
