% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.3-data-validators.R
\name{validate_equation_params}
\alias{validate_equation_params}
\title{Validate equation parameters against requirements}
\usage{
validate_equation_params(category, equation_num, params)
}
\arguments{
\item{category}{Category name}

\item{equation_num}{Equation number as character}

\item{params}{Parameter list}
}
\description{
Validates equation parameters against their category requirements.
}
\keyword{internal}
