% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{validate_fb4_bioenergetic_inputs}
\alias{validate_fb4_bioenergetic_inputs}
\title{Validar entradas para FB4 con objeto Bioenergetic}
\usage{
validate_fb4_bioenergetic_inputs(
  bio_obj,
  fit_to,
  fit_value,
  first_day,
  last_day
)
}
\description{
Validar entradas para FB4 con objeto Bioenergetic
}
\keyword{internal}
