% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{validate_fb4_fit_result}
\alias{validate_fb4_fit_result}
\title{Validar resultado de ajuste FB4}
\usage{
validate_fb4_fit_result(fit_result)
}
\arguments{
\item{fit_result}{Resultado de función de ajuste FB4}
}
\value{
Lista con información de validación
}
\description{
Validar resultado de ajuste FB4
}
