% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{validate_fb4_results}
\alias{validate_fb4_results}
\title{Validar resultados FB4}
\usage{
validate_fb4_results(fb4_result)
}
\arguments{
\item{fb4_result}{Objeto fb4_result}
}
\value{
Lista con resultados de validación
}
\description{
Validar resultados FB4
}
