% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{validate_fb4_results_internal}
\alias{validate_fb4_results_internal}
\title{Función de validación interna para resultados}
\usage{
validate_fb4_results_internal(results)
}
\arguments{
\item{results}{Resultados del modelo}
}
\value{
Lista con validación
}
\description{
Función de validación interna para resultados
}
\keyword{internal}
