% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{validate_fit_result}
\alias{validate_fit_result}
\title{Validar resultado de ajuste}
\usage{
validate_fit_result(fit_result)
}
\arguments{
\item{fit_result}{Resultado de función de ajuste}
}
\value{
Lista con validación
}
\description{
Validar resultado de ajuste
}
