% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.1-strategy-interface.R
\name{validate_fit_to_strategy_concordance}
\alias{validate_fit_to_strategy_concordance}
\title{Validate concordance between fit_to and strategy parameters}
\usage{
validate_fit_to_strategy_concordance(fit_to, strategy)
}
\arguments{
\item{fit_to}{The target fitting parameter}

\item{strategy}{The strategy method to be used}
}
\description{
Ensures that the fit_to parameter matches the expected strategy method.
Throws an error if there's a mismatch.
}
\keyword{internal}
