% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.3-data-validators.R
\name{validate_fitting_settings}
\alias{validate_fitting_settings}
\title{Validate fitting settings}
\usage{
validate_fitting_settings(settings)
}
\arguments{
\item{settings}{Processed simulation settings}
}
\description{
Validates processed simulation settings before execution.
}
\keyword{internal}
