% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{validate_linear_segments_params}
\alias{validate_linear_segments_params}
\title{Validar par치metros de funci칩n lineal por segmentos}
\usage{
validate_linear_segments_params(alpha1, beta1, alpha2, beta2, cutoff)
}
\arguments{
\item{alpha1}{Intercepto primer segmento}

\item{beta1}{Pendiente primer segmento}

\item{alpha2}{Intercepto segundo segmento}

\item{beta2}{Pendiente segundo segmento}

\item{cutoff}{Peso de corte}
}
\value{
Lista con validaci칩n y advertencias
}
\description{
Verifica continuidad y validez de par치metros
}
