% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.1-plot-core.R
\name{validate_plot_data}
\alias{validate_plot_data}
\title{Validate Plot Data Availability}
\usage{
validate_plot_data(daily_data, required_columns, plot_type = "plot")
}
\arguments{
\item{daily_data}{Daily output data frame}

\item{required_columns}{Character vector of required column names}

\item{plot_type}{Name of plot type for error messages}
}
\value{
NULL (throws error if data not available)
}
\description{
Checks if required data columns are available for plotting.
}
\keyword{internal}
