% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{validate_power_function_params}
\alias{validate_power_function_params}
\title{Validar par치metros de funci칩n potencia}
\usage{
validate_power_function_params(alpha1, beta1, weight_range = c(1, 1000))
}
\arguments{
\item{alpha1}{Coeficiente multiplicativo}

\item{beta1}{Exponente}

\item{weight_range}{Rango de pesos para probar (vector de 2 elementos)}
}
\value{
Lista con validaci칩n y advertencias
}
\description{
Verifica que los par치metros produzcan valores realistas
}
