% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{validate_predator_requirements}
\alias{validate_predator_requirements}
\title{Validate predator requirements (special data handling)}
\usage{
validate_predator_requirements(eq_reqs, predator_params, validation)
}
\description{
Validate predator requirements (special data handling)
}
\keyword{internal}
