% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{validate_respiration_requirements}
\alias{validate_respiration_requirements}
\title{Validate respiration requirements (multiple categories)}
\usage{
validate_respiration_requirements(
  eq_reqs,
  respiration_params,
  species_params,
  validation
)
}
\description{
Validate respiration requirements (multiple categories)
}
\keyword{internal}
