% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.4-main-validators.R
\name{validate_statistical_method_inputs}
\alias{validate_statistical_method_inputs}
\title{Validate statistical strategy inputs}
\usage{
validate_statistical_method_inputs(
  strategy,
  observed_weights,
  covariates = NULL
)
}
\description{
Validate statistical strategy inputs
}
\keyword{internal}
