% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.3-data-validators.R
\name{validate_temporal_data}
\alias{validate_temporal_data}
\title{Validate temporal data}
\usage{
validate_temporal_data(
  temperature,
  diet_matrix,
  energy_matrix,
  indigestible_matrix,
  reproduction_data
)
}
\arguments{
\item{temperature}{Temperature vector}

\item{diet_matrix}{Diet proportion matrix}

\item{energy_matrix}{Prey energy matrix}

\item{indigestible_matrix}{Indigestible fraction matrix}

\item{reproduction_data}{Reproduction vector}
}
\description{
Validates temporal data arrays for simulation readiness.
Your original function from data-processing.R, optimized.
}
\keyword{internal}
