% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15.0-analysis-core.R
\name{analysis-core}
\alias{analysis-core}
\title{Core Analysis Functions for FB4 Results}
\description{
Core functions for extracting and accessing results from FB4 simulations.
These functions provide a unified interface to access results regardless
of the fitting method used (basic, optim, MLE, hierarchical).
}
