% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_activity_factor}
\alias{calculate_activity_factor}
\title{Activity factor calculation (Low-level)}
\usage{
calculate_activity_factor(
  weight,
  temperature,
  RTL,
  ACT,
  RK4,
  BACT,
  RK1,
  RK5,
  RTO
)
}
\arguments{
\item{weight}{Fish weight (g)}

\item{temperature}{Water temperature (°C)}

\item{RTL}{Cutoff temperature for activity relationship change}

\item{ACT}{Activity multiplier}

\item{RK4}{Mass dependence coefficient for swimming speed}

\item{BACT}{Temperature dependence coefficient for swimming speed}

\item{RK1}{Intercept for swimming speed above cutoff temperature}

\item{RK5}{Temperature coefficient for swimming speed}

\item{RTO}{Optimum temperature for respiration}
}
\value{
Activity factor
}
\description{
Implements activity function with temperature component
}
\keyword{internal}
