% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_activity_factor_respiration}
\alias{calculate_activity_factor_respiration}
\title{Calculate activity factor for respiration (Mid-level)}
\usage{
calculate_activity_factor_respiration(
  weight,
  temperature,
  respiration_params,
  activity_params
)
}
\arguments{
\item{weight}{Fish weight (g)}

\item{temperature}{Water temperature (°C)}

\item{respiration_params}{List with respiration parameters}

\item{activity_params}{List with activity parameters}
}
\value{
Activity factor
}
\description{
Coordinates activity calculation based on respiration equation
}
\keyword{internal}
