% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_activity_factor_respiration}
\alias{calculate_activity_factor_respiration}
\title{Calcular factor de actividad para respiración}
\usage{
calculate_activity_factor_respiration(
  weight,
  temperature,
  respiration_params,
  activity_params
)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{temperature}{Temperatura del agua (°C)}

\item{respiration_params}{Lista con parámetros de respiración}

\item{activity_params}{Lista con parámetros de actividad}
}
\value{
Factor de actividad
}
\description{
Calcular factor de actividad para respiración
}
\keyword{internal}
