% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{calculate_consumption_params_eq2}
\alias{calculate_consumption_params_eq2}
\title{Calculate additional parameters for consumption equation 2 (Mid-level)}
\usage{
calculate_consumption_params_eq2(CQ, CTM, CTO)
}
\arguments{
\item{CQ}{Temperature coefficient}

\item{CTM}{Maximum temperature}

\item{CTO}{Optimum temperature}
}
\value{
List with CY, CZ, CX
}
\description{
Calculates derived parameters needed for temperature equation 2
}
