% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{calculate_consumption_params_eq3}
\alias{calculate_consumption_params_eq3}
\title{Calculate additional parameters for consumption equation 3 (Mid-level)}
\usage{
calculate_consumption_params_eq3(CTO, CQ, CTL, CTM, CK1, CK4)
}
\arguments{
\item{CTO}{Optimum temperature}

\item{CQ}{Temperature coefficient}

\item{CTL}{Temperature with reduced rate}

\item{CTM}{Maximum temperature}

\item{CK1}{Small fraction of maximum rate}

\item{CK4}{Reduced fraction of maximum rate}
}
\value{
List with CG1, CG2
}
\description{
Calculates derived parameters needed for temperature equation 3
}
