% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{calculate_contaminant_accumulation}
\alias{calculate_contaminant_accumulation}
\title{Calculate contaminant accumulation (Mid-level - Main function)}
\usage{
calculate_contaminant_accumulation(
  respiration_o2,
  consumption,
  weight,
  temperature,
  current_concentration,
  contaminant_params
)
}
\arguments{
\item{respiration_o2}{Respiration in g O2/g/day}

\item{consumption}{Vector of consumption by prey type (g/day)}

\item{weight}{Fish weight (g)}

\item{temperature}{Water temperature (°C)}

\item{current_concentration}{Current concentration in predator (μg/g)}

\item{contaminant_params}{List with contaminant parameters}
}
\value{
List with contaminant results
}
\description{
Main function for calculating contaminant dynamics
}
