% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{calculate_cumulative_survival}
\alias{calculate_cumulative_survival}
\title{Calculate cumulative survival (Utility)}
\usage{
calculate_cumulative_survival(survival_rates, initial_population = 1)
}
\arguments{
\item{survival_rates}{Vector of daily survival rates}

\item{initial_population}{Initial population}
}
\value{
Vector with surviving population by day
}
\description{
Calculates survival across multiple periods
}
