% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{calculate_daily_growth}
\alias{calculate_daily_growth}
\title{Calcula el peso final del día considerando crecimiento y reproducción}
\usage{
calculate_daily_growth(
  current_weight,
  net_energy,
  spawn_energy,
  predator_energy_density
)
}
\arguments{
\item{current_weight}{Peso inicial del día (g)}

\item{net_energy}{Energía neta disponible (J/g/día)}

\item{spawn_energy}{Energía perdida por reproducción (J)}

\item{predator_energy_density}{Densidad energética del depredador (J/g)}
}
\value{
Lista con peso final y cambio de peso
}
\description{
Calcula el peso final del día considerando crecimiento y reproducción
}
\keyword{internal}
