% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{calculate_daily_metabolism}
\alias{calculate_daily_metabolism}
\title{Calculate egestion, excretion, respiration and SDA for one day}
\usage{
calculate_daily_metabolism(
  consumption_energy,
  current_weight,
  temperature,
  p_value,
  species_params,
  oxycal = 13560,
  diet_proportions = NULL,
  indigestible_fractions = NULL
)
}
\arguments{
\item{consumption_energy}{Energy consumption (J/g/day)}

\item{current_weight}{Current weight (g)}

\item{temperature}{Temperature (°C)}

\item{p_value}{Effective p-value}

\item{species_params}{Species parameters}

\item{oxycal}{Oxycalorific coefficient (J/g O2)}

\item{diet_proportions}{Diet item proportions}

\item{indigestible_fractions}{Indigestible fractions by diet item}
}
\value{
List with metabolic processes
}
\description{
Calculate egestion, excretion, respiration and SDA for one day
}
\keyword{internal}
