% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{calculate_dissolved_fraction}
\alias{calculate_dissolved_fraction}
\title{Calculate dissolved fraction of contaminant (Utility)}
\usage{
calculate_dissolved_fraction(poc_concentration, doc_concentration, kow)
}
\arguments{
\item{poc_concentration}{Particulate organic carbon concentration (kg/L)}

\item{doc_concentration}{Dissolved organic carbon concentration (kg/L)}

\item{kow}{Octanol:water partition coefficient}
}
\value{
Dissolved fraction
}
\description{
Calculates dissolved fraction based on organic carbon according to Arnot & Gobas (2004)
}
