% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_egestion}
\alias{calculate_egestion}
\title{Calcular egestión diaria}
\usage{
calculate_egestion(
  consumption,
  temperature,
  p_value,
  egestion_params,
  indigestible_fraction = 0
)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{temperature}{Temperatura del agua (°C)}

\item{p_value}{Proporción del consumo máximo (p-value)}

\item{egestion_params}{Lista con parámetros de egestión}

\item{indigestible_fraction}{Fracción indigerible (para modelo 3)}
}
\value{
Egestión (J/g)
}
\description{
Calcular egestión diaria
}
