% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{calculate_energy_density}
\alias{calculate_energy_density}
\title{Calcular densidad energética a partir de grasa y proteína}
\usage{
calculate_energy_density(
  fat_content,
  protein_content,
  total_weight,
  fat_energy = 39300,
  protein_energy = 23600
)
}
\arguments{
\item{fat_content}{Contenido de grasa (g)}

\item{protein_content}{Contenido de proteína (g)}

\item{total_weight}{Peso total (g)}

\item{fat_energy}{Energía por gramo de grasa (J/g)}

\item{protein_energy}{Energía por gramo de proteína (J/g)}
}
\value{
Densidad energética (J/g peso húmedo)
}
\description{
Calcular densidad energética a partir de grasa y proteína
}
\keyword{internal}
