% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_excretion}
\alias{calculate_excretion}
\title{Calcular excreción diaria}
\usage{
calculate_excretion(
  consumption,
  egestion,
  temperature,
  p_value,
  excretion_params
)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{egestion}{Egestión (J/g)}

\item{temperature}{Temperatura del agua (°C)}

\item{p_value}{Proporción del consumo máximo (p-value)}

\item{excretion_params}{Lista con parámetros de excreción}
}
\value{
Excreción (J/g)
}
\description{
Calcular excreción diaria
}
