% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{calculate_excretion}
\alias{calculate_excretion}
\title{Calculate daily excretion (Mid-level - Main function)}
\usage{
calculate_excretion(
  consumption,
  egestion,
  temperature,
  p_value,
  excretion_params
)
}
\arguments{
\item{consumption}{Consumption (J/g)}

\item{egestion}{Egestion (J/g)}

\item{temperature}{Water temperature (°C)}

\item{p_value}{Proportion of maximum consumption (p-value)}

\item{excretion_params}{List with excretion parameters}
}
\value{
Excretion (J/g)
}
\description{
Main excretion calculation function called from simulation loop
}
