% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{calculate_fat_by_subtraction}
\alias{calculate_fat_by_subtraction}
\title{Calcular contenido de grasa por sustracción}
\usage{
calculate_fat_by_subtraction(
  total_weight,
  water_content,
  protein_content,
  ash_content
)
}
\arguments{
\item{total_weight}{Peso total húmedo (g)}

\item{water_content}{Contenido de agua (g)}

\item{protein_content}{Contenido de proteína (g)}

\item{ash_content}{Contenido de ceniza (g)}
}
\value{
Contenido de grasa (g)
}
\description{
Calcula gramos de grasa restando agua, proteína y ceniza del peso total
}
\keyword{internal}
