% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-main-growth-model.R
\name{calculate_final_weight}
\alias{calculate_final_weight}
\title{Calcular peso final del día}
\usage{
calculate_final_weight(
  initial_weight,
  energy_gain,
  spawn_energy,
  predator_energy_start,
  day,
  equation,
  params
)
}
\arguments{
\item{initial_weight}{Peso al inicio del día}

\item{energy_gain}{Ganancia de energía neta}

\item{spawn_energy}{Energía perdida por reproducción}

\item{predator_energy_start}{Densidad energética inicial del depredador}

\item{day}{Día actual}

\item{equation}{Ecuación de densidad energética}

\item{params}{Parámetros de densidad energética}
}
\value{
Lista con peso final y densidad energética final
}
\description{
Calcular peso final del día
}
\keyword{internal}
