% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{calculate_fish_water_partition}
\alias{calculate_fish_water_partition}
\title{Calcular coeficiente de partición pez:agua}
\usage{
calculate_fish_water_partition(
  fat_fraction,
  protein_ash_fraction,
  water_fraction,
  kow
)
}
\arguments{
\item{fat_fraction}{Fracción de grasa en el pez}

\item{protein_ash_fraction}{Fracción de proteína + ceniza}

\item{water_fraction}{Fracción de agua}

\item{kow}{Coeficiente de partición octanol:agua}
}
\value{
Coeficiente de partición pez:agua
}
\description{
Calcula Kbw basado en composición corporal y Kow según Arnot & Gobas (2004)
}
