% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{calculate_gill_efficiency}
\alias{calculate_gill_efficiency}
\title{Calcular eficiencia de captación branquial}
\usage{
calculate_gill_efficiency(kow)
}
\arguments{
\item{kow}{Coeficiente de partición octanol:agua}
}
\value{
Eficiencia de captación branquial
}
\description{
Calcula eficiencia branquial basada en Kow según Arnot & Gobas (2004)
}
