% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{calculate_np_ratios}
\alias{calculate_np_ratios}
\title{Calculate N:P ratios for all processes (Utility)}
\usage{
calculate_np_ratios(nitrogen_fluxes, phosphorus_fluxes, ratio_type = "mass")
}
\arguments{
\item{nitrogen_fluxes}{List result from calculate_nitrogen_allocation}

\item{phosphorus_fluxes}{List result from calculate_phosphorus_allocation}

\item{ratio_type}{Type of ratio ("mass" or "molar")}
}
\value{
List with N:P ratios
}
\description{
Calculates molar and mass N:P ratios for consumption, growth, excretion and egestion
}
