% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{calculate_nutrient_balance}
\alias{calculate_nutrient_balance}
\title{Calcular balance de nutrientes}
\usage{
calculate_nutrient_balance(consumption, weight_gain, nutrient_params)
}
\arguments{
\item{consumption}{Vector de consumo por tipo de presa (g/día)}

\item{weight_gain}{Ganancia de peso del depredador (g/día)}

\item{nutrient_params}{Lista con parámetros de nutrientes}
}
\value{
Lista con resultados de nutrientes
}
\description{
Función principal para calcular los flujos de nitrógeno y fósforo
}
