% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{calculate_nutrient_balance}
\alias{calculate_nutrient_balance}
\title{Calculate nutrient balance (Mid-level - Main function)}
\usage{
calculate_nutrient_balance(consumption, weight_gain, nutrient_params)
}
\arguments{
\item{consumption}{Vector of consumption by prey type (g/day)}

\item{weight_gain}{Predator weight gain (g/day)}

\item{nutrient_params}{List with nutrient parameters}
}
\value{
List with nutrient results
}
\description{
Main function for calculating nitrogen and phosphorus fluxes
}
