% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{calculate_nutrient_efficiencies}
\alias{calculate_nutrient_efficiencies}
\title{Calculate nutrient retention efficiencies (Utility)}
\usage{
calculate_nutrient_efficiencies(nitrogen_fluxes, phosphorus_fluxes)
}
\arguments{
\item{nitrogen_fluxes}{List result from calculate_nitrogen_allocation}

\item{phosphorus_fluxes}{List result from calculate_phosphorus_allocation}
}
\value{
List with calculated efficiencies
}
\description{
Calculates assimilation and retention efficiencies for N and P
}
