% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{calculate_predator_energy_density}
\alias{calculate_predator_energy_density}
\title{Calcular densidad energética del depredador}
\usage{
calculate_predator_energy_density(
  weight,
  day = 1,
  predator_params,
  energy_data = NULL
)
}
\arguments{
\item{weight}{Peso del pez (g)}

\item{day}{Día de la simulación (para ecuación 1)}

\item{predator_params}{Lista con parámetros de densidad energética}

\item{energy_data}{Vector con datos de densidad energética por día (para ecuación 1)}
}
\value{
Densidad energética (J/g)
}
\description{
Calcular densidad energética del depredador
}
