% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-body-composition.R
\name{calculate_protein_from_water}
\alias{calculate_protein_from_water}
\title{Estimar contenido de proteína a partir del agua}
\usage{
calculate_protein_from_water(water_content)
}
\arguments{
\item{water_content}{Contenido de agua (g)}
}
\value{
Contenido de proteína (g)
}
\description{
Estima gramos de proteína basado en el contenido de agua usando la
regresión de Breck (2014)
}
\references{
Breck, J.E. 2014. Body composition in fishes: body size matters. Aquaculture 433:40-49.
}
\keyword{internal}
