% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_respiration}
\alias{calculate_respiration}
\title{Calculate daily respiration (Mid-level - Main function)}
\usage{
calculate_respiration(temperature, weight, respiration_params, activity_params)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{weight}{Fish weight (g)}

\item{respiration_params}{List with respiration parameters}

\item{activity_params}{List with activity parameters}
}
\value{
Respiration (g O2/g fish/day)
}
\description{
Main respiration calculation function called from simulation loop
}
