% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_respiration}
\alias{calculate_respiration}
\title{Calcular respiración diaria}
\usage{
calculate_respiration(temperature, weight, respiration_params, activity_params)
}
\arguments{
\item{temperature}{Temperatura del agua (°C)}

\item{weight}{Peso del pez (g)}

\item{respiration_params}{Lista con parámetros de respiración}

\item{activity_params}{Lista con parámetros de actividad}
}
\value{
Respiración (g O2/g pez/día)
}
\description{
Función principal que calcula la respiración basada en temperatura y peso
}
