% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_respiration_params_eq2}
\alias{calculate_respiration_params_eq2}
\title{Calcular parámetros adicionales para ecuación de respiración 2}
\usage{
calculate_respiration_params_eq2(RQ, RTM, RTO)
}
\arguments{
\item{RQ}{Tasa Q10}

\item{RTM}{Temperatura máxima}

\item{RTO}{Temperatura óptima}
}
\value{
Lista con RY, RZ, RX
}
\description{
Calcular parámetros adicionales para ecuación de respiración 2
}
