% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_sda}
\alias{calculate_sda}
\title{Calculate Specific Dynamic Action (SDA) (Utility)}
\usage{
calculate_sda(consumption_energy, egestion_energy, SDA_coeff)
}
\arguments{
\item{consumption_energy}{Consumption in energy (J/g)}

\item{egestion_energy}{Egestion in energy (J/g)}

\item{SDA_coeff}{Specific dynamic action coefficient}
}
\value{
SDA in energy (J/g)
}
\description{
Calculates metabolic cost of feeding and digestion
}
