% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{calculate_temperature_factor_consumption}
\alias{calculate_temperature_factor_consumption}
\title{Calcular factor de temperatura para consumo}
\usage{
calculate_temperature_factor_consumption(temperature, consumption_params)
}
\arguments{
\item{temperature}{Temperatura del agua (°C)}

\item{consumption_params}{Lista con parámetros de consumo}
}
\value{
Factor de temperatura
}
\description{
Calcular factor de temperatura para consumo
}
\keyword{internal}
