% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_temperature_factor_respiration}
\alias{calculate_temperature_factor_respiration}
\title{Calculate temperature factor for respiration (Mid-level)}
\usage{
calculate_temperature_factor_respiration(temperature, respiration_params)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{respiration_params}{List with respiration parameters}
}
\value{
Temperature factor
}
\description{
Coordinates temperature equation selection and calculation
}
\keyword{internal}
