% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{calculate_temperature_factor_respiration}
\alias{calculate_temperature_factor_respiration}
\title{Calcular factor de temperatura para respiración}
\usage{
calculate_temperature_factor_respiration(temperature, respiration_params)
}
\arguments{
\item{temperature}{Temperatura del agua (°C)}

\item{respiration_params}{Lista con parámetros de respiración}
}
\value{
Factor de temperatura
}
\description{
Calcular factor de temperatura para respiración
}
\keyword{internal}
