% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{check_data_compatibility}
\alias{check_data_compatibility}
\title{Verificar compatibilidad entre datasets}
\usage{
check_data_compatibility(
  temperature_data,
  diet_data = NULL,
  prey_energy_data = NULL
)
}
\description{
Verificar compatibilidad entre datasets
}
\keyword{internal}
