% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{check_numeric_value}
\alias{check_numeric_value}
\title{Validar entrada numérica}
\usage{
check_numeric_value(value, name, min_val = -Inf, max_val = Inf)
}
\arguments{
\item{value}{Valor a validar}

\item{name}{Nombre del parámetro}

\item{min_val}{Valor mínimo}

\item{max_val}{Valor máximo}
}
\value{
Valor validado
}
\description{
Validar entrada numérica
}
\keyword{internal}
