% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{clamp}
\alias{clamp}
\title{Limitar valor a un rango}
\usage{
clamp(x, min_val, max_val)
}
\arguments{
\item{x}{Valor de entrada}

\item{min_val}{Valor mínimo}

\item{max_val}{Valor máximo}
}
\value{
Valor limitado al rango
}
\description{
Limitar valor a un rango
}
\keyword{internal}
