% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{compare_with_redfield}
\alias{compare_with_redfield}
\title{Compare with Redfield ratios}
\usage{
compare_with_redfield(np_ratios)
}
\arguments{
\item{np_ratios}{List result from calculate_np_ratios}
}
\value{
Data frame with comparison
}
\description{
Compares calculated N:P ratios with Redfield ratio
}
