% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-nutrient-regeneration.R
\name{compare_with_redfield}
\alias{compare_with_redfield}
\title{Comparar con ratios de Redfield}
\usage{
compare_with_redfield(np_ratios)
}
\arguments{
\item{np_ratios}{Lista resultado de calculate_np_ratios}
}
\value{
Data frame con comparación
}
\description{
Compara ratios N:P calculados con el ratio de Redfield
}
