% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{consumption_temp_eq1}
\alias{consumption_temp_eq1}
\title{Funci처n de temperatura para consumo - Ecuaci처n 1}
\usage{
consumption_temp_eq1(temperature, CQ)
}
\arguments{
\item{temperature}{Temperatura del agua (째C)}

\item{CQ}{Coeficiente de temperatura para consumo}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuaci처n de temperatura 1 (exponencial simple)
}
\keyword{internal}
