% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{consumption_temp_eq2}
\alias{consumption_temp_eq2}
\title{Función de temperatura para consumo - Ecuación 2}
\usage{
consumption_temp_eq2(temperature, CTM, CTO, CX)
}
\arguments{
\item{temperature}{Temperatura del agua (°C)}

\item{CTM}{Temperatura máxima donde cesa el consumo}

\item{CTO}{Temperatura óptima de laboratorio}

\item{CX}{Parámetro calculado}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuación de temperatura 2 (Kitchell et al. 1977)
}
\keyword{internal}
