% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{consumption_temp_eq2}
\alias{consumption_temp_eq2}
\title{Temperature function for consumption - Equation 2 (Low-level)}
\usage{
consumption_temp_eq2(temperature, CTM, CTO, CX)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{CTM}{Maximum temperature where consumption ceases}

\item{CTO}{Laboratory optimum temperature}

\item{CX}{Calculated parameter}
}
\value{
Temperature factor
}
\description{
Implements temperature equation 2 (Kitchell et al. 1977)
}
\keyword{internal}
