% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{consumption_temp_eq3}
\alias{consumption_temp_eq3}
\title{Temperature function for consumption - Equation 3 (Low-level)}
\usage{
consumption_temp_eq3(temperature, CQ, CG1, CK1, CG2, CTL, CK4)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{CQ}{Temperature coefficient}

\item{CG1}{Calculated parameter 1}

\item{CK1}{Small fraction of maximum rate}

\item{CG2}{Calculated parameter 2}

\item{CTL}{Temperature with reduced rate}

\item{CK4}{Reduced fraction of maximum rate}
}
\value{
Temperature factor
}
\description{
Implements temperature equation 3 (Thornton and Lessem 1978)
}
\keyword{internal}
