% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-consumption.R
\name{consumption_temp_eq4}
\alias{consumption_temp_eq4}
\title{Temperature function for consumption - Equation 4 (Low-level)}
\usage{
consumption_temp_eq4(temperature, CQ, CK1, CK4)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{CQ}{Linear coefficient}

\item{CK1}{Quadratic coefficient}

\item{CK4}{Cubic coefficient}
}
\value{
Temperature factor
}
\description{
Implements temperature equation 4 (polynomial)
}
\keyword{internal}
