% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{contaminant-accumulation}
\alias{contaminant-accumulation}
\title{Funciones de Acumulación de Contaminantes para el Modelo FB4}
\description{
Funciones de Acumulación de Contaminantes para el Modelo FB4
}
