% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{contaminant_model_2}
\alias{contaminant_model_2}
\title{Modelo de contaminantes 2 - Con eliminación}
\usage{
contaminant_model_2(C, W, Temperature, X_Prey, X_ae, Burden)
}
\arguments{
\item{C}{Vector de consumo por presa (g/día)}

\item{W}{Peso del pez (g)}

\item{Temperature}{Temperatura del agua (°C)}

\item{X_Prey}{Vector de concentraciones en presas (μg/g)}

\item{X_ae}{Vector de eficiencias de asimilación}

\item{Burden}{Carga corporal actual (μg)}
}
\value{
Vector con c(Clearance, Uptake, Burden, NA)
}
\description{
Modelo con captación de alimento y eliminación dependiente de temperatura y peso
Basado en Trudel & Rasmussen (1997) para MeHg
}
