% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{contaminant_model_3}
\alias{contaminant_model_3}
\title{Contaminant model 3 - Arnot & Gobas (2004) (Low-level)}
\usage{
contaminant_model_3(
  respiration_o2,
  consumption,
  weight,
  temperature,
  prey_concentrations,
  assimilation_efficiency,
  current_burden,
  gill_efficiency,
  fish_water_partition,
  water_concentration,
  dissolved_fraction,
  do_saturation
)
}
\arguments{
\item{respiration_o2}{Respiration (g O2/g/day)}

\item{consumption}{Vector of consumption by prey (g/day)}

\item{weight}{Fish weight (g)}

\item{temperature}{Water temperature (°C)}

\item{prey_concentrations}{Vector of concentrations in prey (μg/g)}

\item{assimilation_efficiency}{Vector of assimilation efficiencies}

\item{current_burden}{Current body burden (μg)}

\item{gill_efficiency}{Gill uptake efficiency}

\item{fish_water_partition}{Fish:water partition coefficient}

\item{water_concentration}{Total concentration in water (mg/L)}

\item{dissolved_fraction}{Dissolved fraction}

\item{do_saturation}{Dissolved oxygen saturation (fraction)}
}
\value{
List with clearance, uptake, and new burden
}
\description{
Complete model with uptake from water and food, elimination proportional to respiration
}
\keyword{internal}
