% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion_model_1}
\alias{egestion_model_1}
\title{Egestion model 1 - Basic (Low-level)}
\usage{
egestion_model_1(consumption, FA)
}
\arguments{
\item{consumption}{Consumption (J/g)}

\item{FA}{Egestion fraction}
}
\value{
Egestion (J/g)
}
\description{
Simple egestion model with constant fraction of consumption
}
\keyword{internal}
