% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion_model_2}
\alias{egestion_model_2}
\title{Modelo de egestión 2 - Elliott (1976)}
\usage{
egestion_model_2(consumption, temperature, p_value, FA, FB, FG)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{temperature}{Temperatura del agua (°C)}

\item{p_value}{Proporción del consumo máximo (p-value)}

\item{FA}{Parámetro base de egestión}

\item{FB}{Coeficiente de dependencia de temperatura}

\item{FG}{Coeficiente de dependencia del nivel de alimentación}
}
\value{
Egestión (J/g)
}
\description{
Modelo de egestión dependiente de temperatura y nivel de alimentación
}
\keyword{internal}
