% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion_model_3}
\alias{egestion_model_3}
\title{Modelo de egestión 3 - Stewart et al. (1983)}
\usage{
egestion_model_3(
  consumption,
  temperature,
  p_value,
  FA,
  FB,
  FG,
  indigestible_fraction = 0
)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{temperature}{Temperatura del agua (°C)}

\item{p_value}{Proporción del consumo máximo (p-value)}

\item{FA}{Parámetro base de egestión}

\item{FB}{Coeficiente de dependencia de temperatura}

\item{FG}{Coeficiente de dependencia del nivel de alimentación}

\item{indigestible_fraction}{Fracción indigerible de las presas}
}
\value{
Egestión (J/g)
}
\description{
Modelo que incluye presas indigeribles
}
\keyword{internal}
