% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{egestion_model_4}
\alias{egestion_model_4}
\title{Modelo de egestión 4 - Elliott (1976) sin p-value}
\usage{
egestion_model_4(consumption, temperature, FA, FB)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{temperature}{Temperatura del agua (°C)}

\item{FA}{Parámetro base de egestión}

\item{FB}{Coeficiente de dependencia de temperatura}
}
\value{
Egestión (J/g)
}
\description{
Modelo simplificado sin dependencia del nivel de alimentación
}
\keyword{internal}
