% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{excretion_model_1}
\alias{excretion_model_1}
\title{Modelo de excreción 1 - Básico}
\usage{
excretion_model_1(consumption, egestion, UA)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{egestion}{Egestión (J/g)}

\item{UA}{Fracción de excreción}
}
\value{
Excreción (J/g)
}
\description{
Modelo simple de excreción proporcional a la materia asimilada
}
\keyword{internal}
