% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-egestion-excretion.R
\name{excretion_model_4}
\alias{excretion_model_4}
\title{Modelo de excreción 4 - Sin dependencia de alimentación}
\usage{
excretion_model_4(consumption, egestion, temperature, UA, UB)
}
\arguments{
\item{consumption}{Consumo (J/g)}

\item{egestion}{Egestión (J/g)}

\item{temperature}{Temperatura del agua (°C)}

\item{UA}{Parámetro base de excreción}

\item{UB}{Coeficiente de dependencia de temperatura}
}
\value{
Excreción (J/g)
}
\description{
Modelo de excreción 4 - Sin dependencia de alimentación
}
\keyword{internal}
