% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{export_fb4_results}
\alias{export_fb4_results}
\title{Export results to CSV}
\usage{
export_fb4_results(fb4_result, file_path)
}
\arguments{
\item{fb4_result}{FB4 result object}

\item{file_path}{Path for CSV file}
}
\description{
Export results to CSV
}
