% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{extract_species_parameters}
\alias{extract_species_parameters}
\title{Extraer parámetros de especie}
\usage{
extract_species_parameters(parms, species_num)
}
\description{
Extraer parámetros de especie
}
\keyword{internal}
