% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-plot-fb4.R
\name{fb4-plots}
\alias{fb4-plots}
\title{FB4 Plotting Functions}
\description{
Main plotting functions for FB4 bioenergetic model results.
Provides S3 methods and plot types for fb4_result and Bioenergetic objects.
}
