% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{fit_fb4_binary_search}
\alias{fit_fb4_binary_search}
\title{Binary Search Algorithm for FB4 Fitting}
\usage{
fit_fb4_binary_search(
  species_params,
  initial_weight,
  fit_to,
  fit_value,
  processed_data,
  model_options,
  oxycal = 13560,
  output_daily = TRUE,
  tolerance = 0.001,
  max_iterations = 25
)
}
\arguments{
\item{species_params}{Extracted species parameters}

\item{initial_weight}{Initial weight (g)}

\item{fit_to}{Type of fitting ("weight" or "consumption")}

\item{fit_value}{Target value}

\item{processed_data}{Processed environmental data}

\item{model_options}{Model options}

\item{oxycal}{Oxycalorific coefficient}

\item{output_daily}{Return daily output}

\item{tolerance}{Tolerance for convergence}

\item{max_iterations}{Maximum number of iterations}
}
\value{
List with fitting result and final simulation
}
\description{
Robust implementation of binary search algorithm based on critical functions
}
