% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{get_daily_value}
\alias{get_daily_value}
\title{Obtener valor diario de datos temporales}
\usage{
get_daily_value(data_vector, day)
}
\description{
Obtener valor diario de datos temporales
}
\keyword{internal}
