% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0-bioenergetic-classes.R
\name{get_parameter_value}
\alias{get_parameter_value}
\title{Get Parameter Value from Species Parameters}
\usage{
get_parameter_value(params, param)
}
\arguments{
\item{params}{Species parameters list}

\item{param}{Parameter name to retrieve}
}
\value{
Parameter value or NULL if not found
}
\description{
Retrieves a specific parameter value from species parameter lists,
searching across all parameter categories.
}
