% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{get_parameter_value}
\alias{get_parameter_value}
\title{Obtener valor de parámetro}
\usage{
get_parameter_value(species_params, parameter_name)
}
\description{
Obtener valor de parámetro
}
\keyword{internal}
