% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{get_results}
\alias{get_results}
\title{Get Results from Fitted Model}
\usage{
get_results(x, component = "all")
}
\arguments{
\item{x}{Fitted model object}

\item{component}{Component to extract (default "all")}
}
\value{
Model results or specified component
}
\description{
Generic function for extracting results from fitted bioenergetic models.
}
