% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Operador de null coalescing}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Primer valor}

\item{y}{Valor por defecto si x es NULL}
}
\value{
x si no es NULL, y en caso contrario
}
\description{
Operador de null coalescing
}
