% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-fitting-algorithms.R
\name{\%||\%}
\alias{\%||\%}
\title{Operador %||% para valores por defecto}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Valor a probar}

\item{y}{Valor por defecto si x es NULL}
}
\value{
x si no es NULL, y en caso contrario
}
\description{
Operador %||% para valores por defecto
}
\keyword{internal}
