% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R, R/14-bioenergetic-classes.R
\name{\%||\%}
\alias{\%||\%}
\title{Operador de null coalescing}
\usage{
a \%||\% b

a \%||\% b
}
\arguments{
\item{x}{Primer valor}

\item{y}{Valor por defecto si x es NULL}
}
\value{
x si no es NULL, y en caso contrario
}
\description{
Operador de null coalescing

Operador de valor por defecto
}
\keyword{internal}
