% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{interpolate_time_series}
\alias{interpolate_time_series}
\title{Interpolar series de tiempo}
\usage{
interpolate_time_series(data, value_col, target_days, method = "linear")
}
\arguments{
\item{data}{Data frame con datos}

\item{value_col}{Nombre de la columna con valores}

\item{target_days}{Vector de días objetivo}

\item{method}{Método de interpolación ("linear", "constant")}
}
\value{
Vector interpolado
}
\description{
Función auxiliar para interpolar datos de series de tiempo
}
\keyword{internal}
