% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{is.Bioenergetic}
\alias{is.Bioenergetic}
\title{Test if Object is Bioenergetic}
\usage{
is.Bioenergetic(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating whether object is of class Bioenergetic
}
\description{
Tests whether an object inherits from the Bioenergetic class.
}
\examples{
\dontrun{
bio_obj <- new_bioenergetic("Salmo salar", 10)
is.Bioenergetic(bio_obj)  # TRUE
is.Bioenergetic(list())   # FALSE
}
}
