% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{is.Bioenergetic}
\alias{is.Bioenergetic}
\title{Verificar si un objeto es de clase Bioenergetic}
\usage{
is.Bioenergetic(x)
}
\arguments{
\item{x}{Objeto a verificar}
}
\value{
TRUE si es Bioenergetic, FALSE si no
}
\description{
Verificar si un objeto es de clase Bioenergetic
}
