% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{mortality-reproduction}
\alias{mortality-reproduction}
\title{Funciones de Mortalidad y Reproducción para el Modelo FB4}
\description{
Funciones de Mortalidad y Reproducción para el Modelo FB4
}
