% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{new_bioenergetic}
\alias{new_bioenergetic}
\title{Constructor simplificado para modelos bioenergéticos}
\usage{
new_bioenergetic(
  species,
  initial_weight,
  life_stage = NULL,
  duration = 365,
  temperature_avg = 20,
  temperature_variation = 5,
  diet_composition = list(Zooplankton = 1),
  ...
)
}
\arguments{
\item{species}{Nombre científico o índice de especie}

\item{initial_weight}{Peso inicial (g)}

\item{life_stage}{Estadio de vida}

\item{duration}{Duración (días)}

\item{temperature_avg}{Temperatura promedio (°C)}

\item{temperature_variation}{Variación (°C)}

\item{diet_composition}{Composición de dieta}

\item{...}{Argumentos adicionales}
}
\value{
Objeto Bioenergetic
}
\description{
Constructor simplificado para modelos bioenergéticos
}
