% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{plot.Bioenergetic}
\alias{plot.Bioenergetic}
\title{Método plot para objetos Bioenergetic}
\usage{
\method{plot}{Bioenergetic}(x, type = "growth", ...)
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{type}{Tipo de gráfico}

\item{...}{Argumentos adicionales}
}
\description{
Método plot para objetos Bioenergetic
}
