% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot.fb4_result}
\alias{plot.fb4_result}
\title{Método plot para objetos fb4_result}
\usage{
\method{plot}{fb4_result}(x, type = "growth", ...)
}
\arguments{
\item{x}{Objeto fb4_result}

\item{type}{Tipo de gráfico ("growth", "consumption", "energy", "population")}

\item{...}{Argumentos adicionales para plot}
}
\value{
NULL (genera gráfico)
}
\description{
Método plot para objetos fb4_result
}
