% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot.fb4_result}
\alias{plot.fb4_result}
\title{Main plotting function for FB4 results}
\usage{
\method{plot}{fb4_result}(x, plot_type = "growth", save_plot = NULL, ...)
}
\arguments{
\item{x}{Object of class fb4_result}

\item{plot_type}{Type of plot: "growth", "consumption", "temperature", "energy", "dashboard"}

\item{save_plot}{Path to save plot (optional)}

\item{...}{Additional arguments}
}
\description{
Main plotting function for FB4 results
}
