% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{plot_dashboard}
\alias{plot_dashboard}
\title{Create simple dashboard with key plots}
\usage{
plot_dashboard(fb4_result)
}
\arguments{
\item{fb4_result}{FB4 result object}
}
\description{
Create simple dashboard with key plots
}
