% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.3-plot-analysis.R
\name{plot_distributions.fb4_result}
\alias{plot_distributions.fb4_result}
\title{Plot parameter distributions for bootstrap and hierarchical methods}
\usage{
plot_distributions.fb4_result(
  fb4_result,
  color_scheme = "green",
  show_individuals = TRUE
)
}
\arguments{
\item{fb4_result}{FB4 result object with distribution data}

\item{color_scheme}{Color scheme to use, default "green"}

\item{show_individuals}{For hierarchical: show individual estimates, default TRUE}
}
\value{
NULL (creates plot)
}
\description{
Shows distributions of parameters from bootstrap samples or 
hierarchical individual estimates.
}
\examples{
\dontrun{
plot_distributions.fb4_result(bootstrap_result)
plot_distributions.fb4_result(hierarchical_result)
}
}
