% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.3-plot-analysis.R
\name{plot_growth_temperature_sensitivity}
\alias{plot_growth_temperature_sensitivity}
\title{Plot sensitivity analysis}
\usage{
plot_growth_temperature_sensitivity(
  sensitivity_data,
  temperatures = seq(5, 20, by = 2),
  feeding_levels = c(0.5, 0.75, 1),
  species = NULL,
  ylim = NULL,
  xlim = NULL,
  colors = "grayscale",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sensitivity_data}{Data frame from \code{analyze_growth_temperature_sensitivity}}

\item{temperatures}{Temperature values to test}

\item{feeding_levels}{Feeding levels to test}

\item{species}{Optional species name for plot title}

\item{ylim}{Optional y-axis limits}

\item{xlim}{Optional x-axis limits}

\item{colors}{Color scheme}

\item{verbose}{Show progress messages}

\item{...}{Additional arguments}
}
\value{
NULL (creates plot)
}
\description{
Creates sensitivity analysis plots for temperature and feeding effects.
}
