% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.3-plot-analysis.R
\name{plot_growth_temperature_sensitivity}
\alias{plot_growth_temperature_sensitivity}
\title{Plot sensitivity analysis}
\usage{
plot_growth_temperature_sensitivity(
  sensitivity_data,
  temperatures = seq(5, 20, by = 2),
  feeding_levels = c(0.5, 0.75, 1),
  species = NULL,
  ylim = NULL,
  xlim = NULL,
  colors = "grayscale",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sensitivity_data}{Data frame from \code{analyze_growth_temperature_sensitivity}}

\item{temperatures}{Temperature values to test}

\item{feeding_levels}{Feeding levels to test}

\item{species}{Optional species name for plot title}

\item{ylim}{Optional y-axis limits}

\item{xlim}{Optional x-axis limits}

\item{colors}{Color scheme}

\item{verbose}{Show progress messages}

\item{...}{Additional arguments}
}
\value{
Called for its plotting side-effect. Invisibly returns \code{NULL}.
}
\description{
Creates sensitivity analysis plots for temperature and feeding effects.
}
\examples{
\donttest{
data(fish4_parameters)
sp   <- fish4_parameters[["Oncorhynchus tshawytscha"]]$life_stages$adult
info <- fish4_parameters[["Oncorhynchus tshawytscha"]]$species_info
bio  <- Bioenergetic(
  species_params     = sp,
  species_info       = info,
  environmental_data = list(
    temperature = data.frame(Day = 1:30, Temperature = rep(12, 30))
  ),
  diet_data = list(
    proportions = data.frame(Day = 1:30, Prey1 = 1.0),
    energies    = data.frame(Day = 1:30, Prey1 = 5000),
    prey_names  = "Prey1"
  ),
  simulation_settings = list(initial_weight = 100, duration = 30)
)
bio$species_params$predator$ED_ini <- 5000
bio$species_params$predator$ED_end <- 5500
sens_data <- analyze_growth_temperature_sensitivity(
  bio_obj         = bio,
  temperatures    = c(10, 14),
  p_values        = c(0.4, 0.7),
  simulation_days = 30,
  verbose         = FALSE
)
plot_growth_temperature_sensitivity(sens_data, species = "Chinook")
}
}
