% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16.3-plot-analysis.R
\name{plot_uncertainty.fb4_result}
\alias{plot_uncertainty.fb4_result}
\title{Plot parameter uncertainty for probabilistic methods}
\usage{
plot_uncertainty.fb4_result(
  fb4_result,
  parameters = "all",
  color_scheme = "blue",
  add_ci_text = TRUE
)
}
\arguments{
\item{fb4_result}{FB4 result object with uncertainty estimates}

\item{parameters}{Parameters to plot: "p_value", "consumption", "all", default "all"}

\item{color_scheme}{Color scheme to use, default "blue"}

\item{add_ci_text}{Add confidence interval text, default TRUE}
}
\value{
NULL (creates plot)
}
\description{
Creates plots showing parameter estimates with confidence intervals.
Adapts automatically to the method used (MLE, bootstrap, hierarchical).
}
\examples{
\donttest{
data(fish4_parameters)
sp   <- fish4_parameters[["Oncorhynchus tshawytscha"]]$life_stages$adult
info <- fish4_parameters[["Oncorhynchus tshawytscha"]]$species_info
bio  <- Bioenergetic(
  species_params     = sp,
  species_info       = info,
  environmental_data = list(
    temperature = data.frame(Day = 1:30, Temperature = rep(12, 30))
  ),
  diet_data = list(
    proportions = data.frame(Day = 1:30, Prey1 = 1.0),
    energies    = data.frame(Day = 1:30, Prey1 = 5000),
    prey_names  = "Prey1"
  ),
  simulation_settings = list(initial_weight = 100, duration = 30)
)
bio$species_params$predator$ED_ini <- 5000
bio$species_params$predator$ED_end <- 5500
set.seed(42)
obs_weights <- rnorm(10, mean = 90, sd = 5)
result_mle <- run_fb4(bio, strategy = "mle", fit_to = "Weight",
                      observed_weights = obs_weights, verbose = FALSE)
plot_uncertainty.fb4_result(result_mle)
}
}
