% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{predator-energy-density}
\alias{predator-energy-density}
\title{Funciones de Densidad Energética del Depredador para el Modelo FB4}
\description{
Funciones de Densidad Energética del Depredador para el Modelo FB4
}
