% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{predator_energy_eq1}
\alias{predator_energy_eq1}
\title{Energy density from interpolated data - Equation 1 (Low-level)}
\usage{
predator_energy_eq1(weight, day, energy_data)
}
\arguments{
\item{weight}{Fish weight (g) - not used in this equation}

\item{day}{Simulation day}

\item{energy_data}{Vector with energy densities by day}
}
\value{
Energy density (J/g)
}
\description{
Retrieves energy density from pre-calculated daily data
}
\keyword{internal}
