% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{predator_energy_eq1}
\alias{predator_energy_eq1}
\title{Densidad energética desde datos interpolados (Ecuación 1)}
\usage{
predator_energy_eq1(weight, day, energy_data)
}
\arguments{
\item{weight}{Peso del pez (g) - no usado en esta ecuación}

\item{day}{Día de la simulación}

\item{energy_data}{Vector con densidades energéticas por día}
}
\value{
Densidad energética (J/g)
}
\description{
Densidad energética desde datos interpolados (Ecuación 1)
}
\keyword{internal}
