% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{predator_energy_eq3}
\alias{predator_energy_eq3}
\title{Power function energy density - Equation 3 (Low-level)}
\usage{
predator_energy_eq3(weight, Alpha1, Beta1)
}
\arguments{
\item{weight}{Fish weight (g)}

\item{Alpha1}{Multiplicative coefficient}

\item{Beta1}{Exponent}
}
\value{
Energy density (J/g)
}
\description{
Power relationship between weight and energy density
}
\keyword{internal}
