% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.1-bioenergetic-methods.R
\name{print.Bioenergetic}
\alias{print.Bioenergetic}
\title{Print Method for Bioenergetic Objects (Improved)}
\usage{
\method{print}{Bioenergetic}(x, ...)
}
\arguments{
\item{x}{Bioenergetic object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for Bioenergetic Objects (Improved)
}
