% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{print.Bioenergetic}
\alias{print.Bioenergetic}
\title{Método print para objetos Bioenergetic}
\usage{
\method{print}{Bioenergetic}(x, ...)
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{...}{Argumentos adicionales}
}
\description{
Método print para objetos Bioenergetic
}
