% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.1-bioenergetic-methods.R
\name{print.fb4_result}
\alias{print.fb4_result}
\title{Print Method for fb4_result Objects (Unified)}
\usage{
\method{print}{fb4_result}(x, ...)
}
\arguments{
\item{x}{fb4_result object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for fb4_result Objects (Unified)
}
