% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{process_bioenergetic_data}
\alias{process_bioenergetic_data}
\title{Procesar datos desde objeto Bioenergetic}
\usage{
process_bioenergetic_data(bio_obj, first_day, last_day)
}
\description{
Procesar datos desde objeto Bioenergetic
}
\keyword{internal}
