% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_consumption_params}
\alias{process_consumption_params}
\title{Process consumption parameters}
\usage{
process_consumption_params(consumption_params)
}
\arguments{
\item{consumption_params}{Raw consumption parameters}
}
\value{
Processed consumption parameters with derived values
}
\description{
Process consumption parameters
}
