% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{process_contaminant_data}
\alias{process_contaminant_data}
\title{Procesar datos de contaminantes}
\usage{
process_contaminant_data(contaminant_data, sim_days, prey_names)
}
\arguments{
\item{contaminant_data}{Lista con datos de contaminantes}

\item{sim_days}{Días de simulación}

\item{prey_names}{Nombres de las presas}
}
\value{
Lista con datos de contaminantes procesados
}
\description{
Procesar datos de contaminantes
}
\keyword{internal}
