% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_contaminant_params}
\alias{process_contaminant_params}
\title{Process contaminant parameters}
\usage{
process_contaminant_params(contaminant_params)
}
\arguments{
\item{contaminant_params}{Raw contaminant parameters}
}
\value{
Processed contaminant parameters with calculated coefficients
}
\description{
Process contaminant parameters
}
