% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_excretion_params}
\alias{process_excretion_params}
\title{Process excretion parameters}
\usage{
process_excretion_params(excretion_params)
}
\arguments{
\item{excretion_params}{Raw excretion parameters}
}
\value{
Processed excretion parameters
}
\description{
Process excretion parameters
}
