% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{process_nutrient_data}
\alias{process_nutrient_data}
\title{Procesar datos de nutrientes}
\usage{
process_nutrient_data(nutrient_data, sim_days, prey_names)
}
\arguments{
\item{nutrient_data}{Lista con datos de nutrientes}

\item{sim_days}{Días de simulación}

\item{prey_names}{Nombres de las presas}
}
\value{
Lista con datos de nutrientes procesados
}
\description{
Procesar datos de nutrientes
}
\keyword{internal}
