% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_nutrient_params}
\alias{process_nutrient_params}
\title{Process nutrient parameters}
\usage{
process_nutrient_params(nutrient_params)
}
\arguments{
\item{nutrient_params}{Raw nutrient parameters}
}
\value{
Processed nutrient parameters
}
\description{
Process nutrient parameters
}
