% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{process_predator_energy_data}
\alias{process_predator_energy_data}
\title{Procesar datos de densidad energética del depredador}
\usage{
process_predator_energy_data(data, first_day, last_day)
}
\arguments{
\item{data}{Data frame con datos de densidad energética}

\item{first_day}{Primer día de simulación}

\item{last_day}{Último día de simulación}
}
\value{
Vector interpolado de densidades energéticas
}
\description{
Prepara y valida datos de densidad energética para interpolación
}
