% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-parameter-processing.R
\name{process_predator_params}
\alias{process_predator_params}
\title{Process predator energy density parameters}
\usage{
process_predator_params(predator_params, n_days = NULL)
}
\arguments{
\item{predator_params}{Raw predator parameters}
}
\value{
Processed predator parameters with energy data
}
\description{
Process predator energy density parameters
}
