% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-data-processing.R
\name{process_simulation_settings}
\alias{process_simulation_settings}
\title{Process simulation settings}
\usage{
process_simulation_settings(settings, first_day, last_day, oxycal = 13560)
}
\arguments{
\item{settings}{Raw simulation settings}

\item{first_day}{First simulation day}

\item{last_day}{Last simulation day}
}
\value{
Processed simulation settings
}
\description{
Process simulation settings
}
