% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{respiration_temp_eq1}
\alias{respiration_temp_eq1}
\title{Temperature function for respiration - Equation 1 (Low-level)}
\usage{
respiration_temp_eq1(temperature, RQ)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{RQ}{Q10 rate for low temperatures}
}
\value{
Temperature factor
}
\description{
Implements temperature equation 1 (simple exponential)
}
\keyword{internal}
