% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{respiration_temp_eq2}
\alias{respiration_temp_eq2}
\title{Temperature function for respiration - Equation 2 (Low-level)}
\usage{
respiration_temp_eq2(temperature, RTM, RTO, RX)
}
\arguments{
\item{temperature}{Water temperature (°C)}

\item{RTM}{Maximum (lethal) temperature}

\item{RTO}{Optimum temperature for respiration}

\item{RX}{Calculated parameter}
}
\value{
Temperature factor
}
\description{
Implements temperature equation 2 (Kitchell et al. 1977)
}
\keyword{internal}
