% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-respiration.R
\name{respiration_temp_eq2}
\alias{respiration_temp_eq2}
\title{Función de temperatura para respiración - Ecuación 2}
\usage{
respiration_temp_eq2(temperature, RTM, RTO, RX)
}
\arguments{
\item{temperature}{Temperatura del agua (°C)}

\item{RTM}{Temperatura máxima (letal)}

\item{RTO}{Temperatura óptima para respiración}

\item{RX}{Parámetro calculado}
}
\value{
Factor de temperatura
}
\description{
Implementa la ecuación de temperatura 2 (Kitchell et al. 1977)
}
\keyword{internal}
