% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4.Bioenergetic}
\alias{run_fb4.Bioenergetic}
\title{Execute Fish Bioenergetics 4.0 Model with Bioenergetic Object}
\usage{
\method{run_fb4}{Bioenergetic}(
  x,
  fit_to = "Weight",
  fit_value,
  first_day = 1,
  last_day = NULL,
  oxycal = 13560,
  output_daily = TRUE,
  tolerance = 0.001,
  max_iterations = 25,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class 'Bioenergetic' with configured parameters and data}

\item{fit_to}{Fitting method: "Weight", "Consumption", "p-value", "Ration", "Ration_prey"}

\item{fit_value}{Target value according to the fitting method}

\item{first_day}{First simulation day (default 1)}

\item{last_day}{Last simulation day (default uses complete duration)}

\item{oxycal}{Oxycalorific coefficient (J/g O2), default 13560}

\item{output_daily}{Logical, whether to include daily output, default TRUE}

\item{tolerance}{Tolerance for iterative fitting, default 0.001}

\item{max_iterations}{Maximum number of iterations, default 25}

\item{verbose}{Logical, whether to show progress messages (default FALSE)}

\item{...}{Additional arguments}
}
\value{
Object of class 'fb4_result' with simulation results
}
\description{
Main function to execute Fish Bioenergetics 4.0 model simulations
using a Bioenergetic class object that already contains all necessary parameters.
}
