% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4.default}
\alias{run_fb4.default}
\title{Default Method for run_fb4}
\usage{
\method{run_fb4}{default}(x, ...)
}
\arguments{
\item{x}{Unsupported object}

\item{...}{Additional arguments}
}
\value{
Throws an error with information about supported classes
}
\description{
Default Method for run_fb4
}
