% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4.default}
\alias{run_fb4.default}
\title{Default Method for run_fb4}
\usage{
\method{run_fb4}{default}(x, ...)
}
\arguments{
\item{x}{Unsupported object}

\item{...}{Additional arguments}
}
\value{
Throws an error with information about supported classes
}
\description{
Default method for run_fb4 generic function. This method is called
when no specific method exists for the given object class.
}
\details{
This method provides helpful error messages when users attempt to use
run_fb4 with unsupported object types, guiding them toward supported
alternatives.
}
\examples{
\dontrun{
# This will throw an error
run_fb4(data.frame(x = 1:10))
}
}
