% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14.0-run-fb4-orchestrator.R
\name{run_fb4}
\alias{run_fb4}
\title{Default run_fb4 method}
\usage{
run_fb4(x, ...)
}
\arguments{
\item{x}{Object to run FB4 simulation on}

\item{...}{Arguments passed to methods}
}
\value{
An object of class \code{fb4_result}. See \code{\link{run_fb4.Bioenergetic}}
  for full details of the return structure.
}
\description{
Generic function that dispatches to appropriate S3 method.
}
