% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4}
\alias{run_fb4}
\title{Generic Function for Fish Bioenergetics 4.0 Model Execution}
\usage{
run_fb4(x, ...)
}
\arguments{
\item{x}{Object to use for simulation (e.g., Bioenergetic object)}

\item{...}{Additional arguments passed to specific methods}
}
\value{
Object of class 'fb4_result' with simulation results
}
\description{
Generic Function for Fish Bioenergetics 4.0 Model Execution
}
