% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{run_fb4_with_ration_grams}
\alias{run_fb4_with_ration_grams}
\title{Execute FB4 with Daily Ration in Grams}
\usage{
run_fb4_with_ration_grams(
  species_params,
  initial_weight,
  ration_grams,
  processed_data,
  model_options,
  oxycal = 13560,
  output_daily = TRUE
)
}
\description{
Execute FB4 with Daily Ration in Grams
}
\keyword{internal}
