% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{safe_exp}
\alias{safe_exp}
\title{Funci칩n segura para exponencial}
\usage{
safe_exp(x, max_exp = 700)
}
\arguments{
\item{x}{Valor de entrada}

\item{max_exp}{Exponente m치ximo permitido (por defecto 700)}
}
\value{
Exponencial o valor m치ximo
}
\description{
Funci칩n segura para exponencial
}
\keyword{internal}
