% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{safe_sqrt}
\alias{safe_sqrt}
\title{Función segura para raíz cuadrada}
\usage{
safe_sqrt(x, min_val = 0)
}
\arguments{
\item{x}{Valor de entrada}

\item{min_val}{Valor mínimo permitido (por defecto 0)}
}
\value{
Raíz cuadrada o error controlado
}
\description{
Función segura para raíz cuadrada
}
\keyword{internal}
