% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_diet}
\alias{set_diet}
\title{Set Diet Data}
\usage{
set_diet(x, diet_proportions, prey_energies)
}
\arguments{
\item{x}{Model object}

\item{diet_proportions}{Diet composition data frame}

\item{prey_energies}{Prey energy density data frame}
}
\value{
Updated model object
}
\description{
Generic function for setting diet data in model objects.
}
