% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_diet}
\alias{set_diet}
\title{Configurar datos de dieta}
\usage{
set_diet(x, diet_proportions, prey_energies)
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{diet_proportions}{Data frame con proporciones de dieta}

\item{prey_energies}{Data frame con energías de presas}
}
\value{
Objeto Bioenergetic modificado
}
\description{
Configurar datos de dieta
}
