% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0-bioenergetic-classes.R
\name{set_environment}
\alias{set_environment}
\title{Set Environmental Data for Bioenergetic Objects}
\usage{
set_environment(x, temperature_data)
}
\arguments{
\item{x}{Bioenergetic object}

\item{temperature_data}{Data frame with Day and Temperature columns}
}
\value{
Updated Bioenergetic object with new environmental data
}
\description{
Updates the environmental data component of a Bioenergetic object
with new temperature information.
}
