% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_environment}
\alias{set_environment}
\title{Set Environmental Data}
\usage{
set_environment(x, temperature_data)
}
\arguments{
\item{x}{Model object}

\item{temperature_data}{Temperature data frame}
}
\value{
Updated model object
}
\description{
Generic function for setting environmental data in model objects.
}
