% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_environment}
\alias{set_environment}
\title{Configurar datos ambientales}
\usage{
set_environment(x, temperature_data, ...)
}
\arguments{
\item{x}{Objeto Bioenergetic}

\item{temperature_data}{Data frame con datos de temperatura}

\item{...}{Datos ambientales adicionales}
}
\value{
Objeto Bioenergetic modificado
}
\description{
Configurar datos ambientales
}
