% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13.0-bioenergetic-classes.R
\name{set_parameter_value}
\alias{set_parameter_value}
\title{Set Parameter Value in Species Parameters}
\usage{
set_parameter_value(params, param, value)
}
\arguments{
\item{params}{Species parameters list}

\item{param}{Parameter name to set}

\item{value}{New parameter value}
}
\value{
Updated parameters list
}
\description{
Sets a specific parameter value in species parameter lists,
automatically finding the correct category.
}
