% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{set_parameter_value}
\alias{set_parameter_value}
\title{Establecer valor de parámetro}
\usage{
set_parameter_value(species_params, parameter_name, value)
}
\description{
Establecer valor de parámetro
}
\keyword{internal}
