% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{setup_model_options}
\alias{setup_model_options}
\title{Configurar opciones del modelo}
\usage{
setup_model_options(
  calc_mortality = FALSE,
  calc_reproduction = FALSE,
  calc_contaminant = FALSE,
  calc_nutrient = FALSE,
  ...
)
}
\arguments{
\item{calc_mortality}{Calcular mortalidad}

\item{calc_reproduction}{Calcular reproducción}

\item{calc_contaminant}{Calcular contaminantes}

\item{calc_nutrient}{Calcular nutrientes}

\item{...}{Opciones adicionales}
}
\value{
Lista con opciones del modelo
}
\description{
Configurar opciones del modelo
}
\keyword{internal}
