% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{solve_weight_iterative}
\alias{solve_weight_iterative}
\title{Resolver peso usando método iterativo (método general seguro)}
\usage{
solve_weight_iterative(
  target_energy,
  predator_params,
  day = 1,
  initial_guess = 1000
)
}
\arguments{
\item{target_energy}{Energía corporal objetivo}

\item{predator_params}{Parámetros del depredador}

\item{day}{Día actual}

\item{initial_guess}{Estimación inicial del peso}
}
\value{
Peso final
}
\description{
Resolver peso usando método iterativo (método general seguro)
}
