% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{solve_weight_linear_segments}
\alias{solve_weight_linear_segments}
\title{Resolver peso para función lineal por segmentos (PREDEDEQ=2)}
\usage{
solve_weight_linear_segments(available_energy, predator_params)
}
\arguments{
\item{available_energy}{Energía disponible (J)}

\item{predator_params}{Parámetros de densidad energética}
}
\value{
Peso final (g)
}
\description{
Resolver peso para función lineal por segmentos (PREDEDEQ=2)
}
