% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{solve_weight_linear_segments}
\alias{solve_weight_linear_segments}
\title{Solve weight for linear segments (Low-level)}
\usage{
solve_weight_linear_segments(available_energy, predator_params)
}
\arguments{
\item{available_energy}{Available energy (J)}

\item{predator_params}{Predator parameters}
}
\value{
Final weight (g)
}
\description{
Solves final weight for piecewise linear function (PREDEDEQ = 2)
}
\keyword{internal}
