% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-predator-energy-density.R
\name{solve_weight_power_function}
\alias{solve_weight_power_function}
\title{Resolver peso final para función potencia (PREDEDEQ = 1 y 3)}
\usage{
solve_weight_power_function(initial_weight, net_energy, Alpha1, Beta1)
}
\arguments{
\item{initial_weight}{Peso inicial}

\item{net_energy}{Energía neta}

\item{Alpha1}{Coeficiente}

\item{Beta1}{Exponente}
}
\value{
Peso final
}
\description{
Resolver peso final para función potencia (PREDEDEQ = 1 y 3)
}
\keyword{internal}
