% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-bioenergetic-classes.R
\name{summary.Bioenergetic}
\alias{summary.Bioenergetic}
\title{Método summary para objetos Bioenergetic}
\usage{
\method{summary}{Bioenergetic}(object, ...)
}
\arguments{
\item{object}{Objeto Bioenergetic}

\item{...}{Argumentos adicionales}
}
\description{
Método summary para objetos Bioenergetic
}
