% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{summary.fb4_result}
\alias{summary.fb4_result}
\title{Summary method for fb4_result objects}
\usage{
\method{summary}{fb4_result}(object, ...)
}
\arguments{
\item{object}{FB4 result object}

\item{...}{Additional arguments}
}
\description{
Summary method for fb4_result objects
}
