% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-visualization-analysis.R
\name{summary.fb4_result}
\alias{summary.fb4_result}
\title{Método summary para objetos fb4_result}
\usage{
\method{summary}{fb4_result}(object, ...)
}
\arguments{
\item{object}{Objeto fb4_result}

\item{...}{Argumentos adicionales}
}
\value{
Resumen del objeto
}
\description{
Método summary para objetos fb4_result
}
