% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_basic_params}
\alias{validate_basic_params}
\title{Validar parámetros básicos del modelo}
\usage{
validate_basic_params(initial_weight, duration)
}
\arguments{
\item{initial_weight}{Peso inicial}

\item{duration}{Duración en días}
}
\value{
NULL (lanza error si inválido)
}
\description{
Validar parámetros básicos del modelo
}
\keyword{internal}
