% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{validate_bioenergetic_for_simulation}
\alias{validate_bioenergetic_for_simulation}
\title{Comprehensive validation for Bioenergetic objects}
\usage{
validate_bioenergetic_for_simulation(bio_obj)
}
\arguments{
\item{bio_obj}{Bioenergetic object}
}
\value{
List with validation results
}
\description{
Comprehensive validation for Bioenergetic objects
}
