% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-contaminant-accumulation.R
\name{validate_contaminant_params}
\alias{validate_contaminant_params}
\title{Validar par치metros de contaminantes}
\usage{
validate_contaminant_params(contaminant_params)
}
\arguments{
\item{contaminant_params}{Lista con par치metros}
}
\value{
Lista con resultados de validaci칩n
}
\description{
Validar par치metros de contaminantes
}
