% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.3-data-validators.R
\name{validate_data_consistency}
\alias{validate_data_consistency}
\title{Validate data consistency between components}
\usage{
validate_data_consistency(simulation_data, n_prey_temporal)
}
\arguments{
\item{simulation_data}{Complete simulation data}

\item{n_prey_temporal}{Number of prey species in temporal data}
}
\description{
Checks cross-component consistency within processed simulation data.
}
\keyword{internal}
