% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_diet_consistency}
\alias{validate_diet_consistency}
\title{Validar consistencia entre datos de dieta y energía}
\usage{
validate_diet_consistency(diet_data, energy_data)
}
\arguments{
\item{diet_data}{Data frame con proporciones de dieta}

\item{energy_data}{Data frame con energías de presas}
}
\value{
NULL (lanza error si inconsistente)
}
\description{
Validar consistencia entre datos de dieta y energía
}
\keyword{internal}
