% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-fb4-main.R
\name{validate_fb4_inputs}
\alias{validate_fb4_inputs}
\title{Validar entradas para FB4}
\usage{
validate_fb4_inputs(
  species,
  initial_weight,
  fit_to,
  fit_value,
  first_day,
  last_day,
  temperature_data,
  diet_data,
  prey_energy_data
)
}
\description{
Validar entradas para FB4
}
\keyword{internal}
