% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{validate_fb4_inputs}
\alias{validate_fb4_inputs}
\title{Validate inputs for FB4 simulation}
\usage{
validate_fb4_inputs(
  bio_obj,
  fit_to = NULL,
  fit_value = NULL,
  first_day = 1,
  last_day = NULL
)
}
\arguments{
\item{bio_obj}{Bioenergetic object}

\item{fit_to}{Fitting target}

\item{fit_value}{Fitting value}

\item{first_day}{First simulation day}

\item{last_day}{Last simulation day}
}
\value{
NULL (throws error if invalid)
}
\description{
Validate inputs for FB4 simulation
}
