% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_fb4_inputs}
\alias{validate_fb4_inputs}
\title{Validar entradas del modelo FB4}
\usage{
validate_fb4_inputs(
  species,
  initial_weight,
  fit_to,
  fit_value,
  first_day,
  last_day,
  temperature_data,
  diet_data,
  prey_energy_data
)
}
\arguments{
\item{species}{Número o nombre de especie}

\item{initial_weight}{Peso inicial}

\item{fit_to}{Tipo de ajuste}

\item{fit_value}{Valor de ajuste}

\item{first_day}{Primer día}

\item{last_day}{Último día}

\item{temperature_data}{Datos de temperatura}

\item{diet_data}{Datos de dieta}

\item{prey_energy_data}{Datos de energía de presas}
}
\value{
NULL (solo valida, detiene si hay errores)
}
\description{
Validar entradas del modelo FB4
}
\keyword{internal}
