% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-mortality-reproduction.R
\name{validate_mortality_reproduction_params}
\alias{validate_mortality_reproduction_params}
\title{Validar parámetros de mortalidad y reproducción}
\usage{
validate_mortality_reproduction_params(
  mortality_rates = NULL,
  reproduction_fractions = NULL,
  species_type = "general"
)
}
\arguments{
\item{mortality_rates}{Vector o matriz de tasas de mortalidad}

\item{reproduction_fractions}{Vector de fracciones reproductivas}

\item{species_type}{Tipo de especie ("small_fish", "large_fish", "general")}
}
\value{
Lista con validación
}
\description{
Verifica que los parámetros estén en rangos realistas
}
