% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-data-processing.R
\name{validate_processed_data}
\alias{validate_processed_data}
\title{Validar consistencia de datos procesados}
\usage{
validate_processed_data(processed_data)
}
\arguments{
\item{processed_data}{Datos procesados}
}
\value{
Lista con resultados de validación
}
\description{
Validar consistencia de datos procesados
}
