% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_respiration_params}
\alias{validate_respiration_params}
\title{Validar parámetros de respiración}
\usage{
validate_respiration_params(respiration_params)
}
\arguments{
\item{respiration_params}{Lista con parámetros de respiración}
}
\value{
Lista con resultados de validación
}
\description{
Validar parámetros de respiración
}
\keyword{internal}
