% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_species_params}
\alias{validate_species_params}
\title{Validar par치metros de especies para FB4}
\usage{
validate_species_params(species_params)
}
\arguments{
\item{species_params}{Lista con par치metros de especies}
}
\value{
Lista con resultados de validaci칩n
}
\description{
Validar par치metros de especies para FB4
}
