% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_species_params}
\alias{validate_species_params}
\title{Validar parámetros de especies para FB4}
\usage{
validate_species_params(species_params)
}
\arguments{
\item{species_params}{Lista con parámetros de especies}
}
\value{
Lista con campos: valid (bool), errors, warnings
}
\description{
Validar parámetros de especies para FB4
}
