% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{validate_time_series_data}
\alias{validate_time_series_data}
\title{Validar datos de series de tiempo}
\usage{
validate_time_series_data(
  data,
  data_name,
  required_cols = NULL,
  min_cols = NULL
)
}
\arguments{
\item{data}{Datos a validar}

\item{data_name}{Nombre del conjunto de datos (para mensajes de error)}

\item{required_cols}{Columnas requeridas}

\item{min_cols}{Número mínimo de columnas}
}
\description{
Validar datos de series de tiempo
}
\keyword{internal}
